% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_AlleleMatrix.R
\docType{class}
\name{AlleleMatrix}
\alias{AlleleMatrix}
\title{AlleleMatrix Class}
\description{
The AlleleMatrix class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the AlleleMatrix entity
}
\details{
The AlleleMatrix class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the AlleleMatrix entity
}
\examples{
AlleleMatrix$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AlleleMatrix-new}{\code{AlleleMatrix$new()}}
\item \href{#method-AlleleMatrix-search}{\code{AlleleMatrix$search()}}
\item \href{#method-AlleleMatrix-searchResult}{\code{AlleleMatrix$searchResult()}}
\item \href{#method-AlleleMatrix-clone}{\code{AlleleMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlleleMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-AlleleMatrix-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlleleMatrix$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlleleMatrix-search"></a>}}
\if{latex}{\out{\hypertarget{method-AlleleMatrix-search}{}}}
\subsection{Method \code{search()}}{
Searches for AlleleMatrix objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlleleMatrix$search(
  pagination = NULL,
  preview = NULL,
  dataMatrixNames = NULL,
  dataMatrixAbbreviations = NULL,
  positionRanges = NULL,
  germplasmNames = NULL,
  germplasmPUIs = NULL,
  germplasmDbIds = NULL,
  sampleDbIds = NULL,
  callSetDbIds = NULL,
  variantDbIds = NULL,
  variantSetDbIds = NULL,
  expandHomozygotes = NULL,
  sepPhased = NULL,
  sepUnphased = NULL,
  unknownString = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pagination}}{Pagination for the matrix}

\item{\code{preview}}{Default Value = false <br/> If 'preview' is set to true, then the server should only return the lists of 'callSetDbIds',  'variantDbIds', and 'variantSetDbIds'.}

\item{\code{dataMatrixNames}}{\code{dataMatrixNames} is a list of names (ie 'Genotype', 'Read Depth' etc).}

\item{\code{dataMatrixAbbreviations}}{\code{dataMatrixAbbreviations} is a comma seperated list of abbreviations (ie 'GT', 'RD' etc).}

\item{\code{positionRanges}}{The postion range to search <br/> Uses the format "<chrom>:<start>-<end>" where <chrom> is the chromosome name, <start> is  the starting position of the range, and <end> is the ending position of the range}

\item{\code{germplasmNames}}{A list of human readable \code{Germplasm} names}

\item{\code{germplasmPUIs}}{A list of permanent unique identifiers associated with \code{Germplasm}}

\item{\code{germplasmDbIds}}{A list of IDs which uniquely identify \code{Germplasm} within the given database server}

\item{\code{sampleDbIds}}{A list of IDs which uniquely identify \code{Samples} within the given database server}

\item{\code{callSetDbIds}}{A list of IDs which uniquely identify \code{CallSets} within the given database server}

\item{\code{variantDbIds}}{A list of IDs which uniquely identify \code{Variants} within the given database server}

\item{\code{variantSetDbIds}}{A list of IDs which uniquely identify \code{VariantSets} within the given database server}

\item{\code{expandHomozygotes}}{Should homozygotes be expanded (true) or collapsed into a single occurrence (false)}

\item{\code{sepPhased}}{The string used as a separator for phased allele calls.}

\item{\code{sepUnphased}}{The string used as a separator for unphased allele calls.}

\item{\code{unknownString}}{The string used as a representation for missing data.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for AlleleMatrix objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlleleMatrix-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-AlleleMatrix-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlleleMatrix$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AlleleMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AlleleMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AlleleMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
