% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\docType{class}
\name{BaseAuth}
\alias{BaseAuth}
\title{BaseAuth Class}
\description{
The BaseAuth is a super class for all classes that provide authentication functionality for httr2 requests
}
\details{
The BaseAuth is a super class for all classes that provide authentication functionality for httr2 requests
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseAuth-new}{\code{BaseAuth$new()}}
\item \href{#method-BaseAuth-authentication}{\code{BaseAuth$authentication()}}
\item \href{#method-BaseAuth-clone}{\code{BaseAuth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAuth-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAuth-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAuth$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAuth-authentication"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAuth-authentication}{}}}
\subsection{Method \code{authentication()}}{
Provides the authentication function for httr2 requests
needs to be implemented in subclasses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAuth$authentication(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{The httr2 request object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The httr2 request object with the Basic Authorization header added
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseAuth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseAuth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseAuth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
