% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrAPIBaseClient.R
\docType{class}
\name{BaseBrAPIClient}
\alias{BaseBrAPIClient}
\title{BrAPIClient Class}
\description{
The BaseBrAPIClient class handles the configuration of the BrAPI connection and
is a wrapper class around httr2 functionality. Do not use directly, only use the generated subclass BrAPIClient
}
\details{
The BaseBrAPIClient class handles the configuration of the BrAPIClient connection and
is a wrapper class around httr2 functionality
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{server}}{Get or sets the server. In the case of setting it returns the BrAPIClient to allow for chaining.}

\item{\code{format}}{Get or sets the format format of the response from the server, either  'list', 'json', 'string' or 'raw'.
In the case of setting it returns the BrAPIClient to allow for chaining.}

\item{\code{authentication}}{Sets the authentication, which can either by a Bearer token or a authentication provider function}

\item{\code{dry_run}}{Get or sets if the client should perform a dry run
In the case of setting it returns the BrAPIClient to allow for chaining.}

\item{\code{verbosity}}{Get or sets the verbosity level to be used. See verbosity in httr2 package for details
In the case of setting it returns the BrAPIClient to allow for chaining.}

\item{\code{max_tries}}{Get or sets the maximum number of attempt each request will make.
In the case of setting it returns the BrAPIClient to allow for chaining.}

\item{\code{multi}}{Controls what happens when a query argument value is a vector. If null it passes the
query argument value as a concatenated string. Eg. c('example1', 'example1') will
be converted to \link{'example1', 'example1'}
see .multi in https://httr2.r-lib.org/reference/req_url.html'
In the case of setting it returns the BrAPIClient to allow for chaining.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseBrAPIClient-new}{\code{BaseBrAPIClient$new()}}
\item \href{#method-BaseBrAPIClient-perform_get_request}{\code{BaseBrAPIClient$perform_get_request()}}
\item \href{#method-BaseBrAPIClient-perform_post_request}{\code{BaseBrAPIClient$perform_post_request()}}
\item \href{#method-BaseBrAPIClient-clone}{\code{BaseBrAPIClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseBrAPIClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseBrAPIClient-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the getBrAPI function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBrAPIClient$new(
  server = NULL,
  format = "list",
  authentication = NULL,
  dry_run = FALSE,
  verbosity = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{server}}{The BraPI server URL to be used}

\item{\code{format}}{The format of the response, either 'list', 'json', 'string' or 'raw'
(default is 'list')
'list' will parse the JSON response into an R list, with 3 elements: data (in tibble), pagination and status.
'json' will return the JSON response as a character string.
'raw' will return the raw response as a raw bytes
'string' will return the response as a character string}

\item{\code{authentication}}{The authentication can either by a Bearer token or an authentication provider function}

\item{\code{dry_run}}{If TRUE the client will perform a dry run and not actually perform the requests, useful for debugging}

\item{\code{verbosity}}{The verbosity level to be used. See verbosity in httr2 package for details}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseBrAPIClient-perform_get_request"></a>}}
\if{latex}{\out{\hypertarget{method-BaseBrAPIClient-perform_get_request}{}}}
\subsection{Method \code{perform_get_request()}}{
Performs a GET request to the BraPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBrAPIClient$perform_get_request(
  endpoint = NULL,
  queryParams = list(),
  page = NULL,
  pageSize = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{The BraPI endpoint to be called}

\item{\code{queryParams}}{The query parameters to be used}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the BraPI server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseBrAPIClient-perform_post_request"></a>}}
\if{latex}{\out{\hypertarget{method-BaseBrAPIClient-perform_post_request}{}}}
\subsection{Method \code{perform_post_request()}}{
Performs a POST request to the BraPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBrAPIClient$perform_post_request(
  endpoint = NULL,
  queryParams = list(),
  page = NULL,
  pageSize = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{The BraPI endpoint to be called}

\item{\code{queryParams}}{The query parameters to be used}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the BraPI server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseBrAPIClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseBrAPIClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseBrAPIClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
