% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\docType{class}
\name{BasicAuth}
\alias{BasicAuth}
\title{BasicAuth Class}
\description{
The BasicAuth class handles Authorization for httr2 requests
}
\details{
The BasicAuth class provides basic authentication functionality for httr2 requests
}
\seealso{
Other authentication-classes: 
\code{\link{BaseAuth}},
\code{\link{OAuthClientCredentials}},
\code{\link{OAuthFlow}}
}
\concept{authentication-classes}
\keyword{authentication-class}
\section{Super class}{
\code{\link[BrAPI.R6:BaseAuth]{BrAPI.R6::BaseAuth}} -> \code{BasicAuth}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasicAuth-new}{\code{BasicAuth$new()}}
\item \href{#method-BasicAuth-authentication}{\code{BasicAuth$authentication()}}
\item \href{#method-BasicAuth-clone}{\code{BasicAuth$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BrAPI.R6" data-topic="BaseAuth" data-id="token"><a href='../../BrAPI.R6/html/BaseAuth.html#method-BaseAuth-token'><code>BrAPI.R6::BaseAuth$token()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the basic_authentication() function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$new(username = NULL, password = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{The username to be used for Authorization.}

\item{\code{password}}{The password to be used for Authorization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-authentication"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-authentication}{}}}
\subsection{Method \code{authentication()}}{
Provides the authentication function for httr2 requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$authentication(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{The httr2 request object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The httr2 request object with the Authorization header added
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
