% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_BrAPIClient.R
\docType{class}
\name{BrAPIClient}
\alias{BrAPIClient}
\title{BrAPIClient Class}
\description{
The BrAPIClient class handles the configuration of the BrAPI connection and
is a wrapper class around httr2 functionality
}
\details{
The BrAPIClient class handles the configuration of the BrAPIClient connection and
is a wrapper class around httr2 functionality
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Super class}{
\code{\link[BrAPI.R6:BaseBrAPIClient]{BrAPI.R6::BaseBrAPIClient}} -> \code{BrAPIClient}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{alleleMatrix}}{Get the AlleleMatrix R6 class object which can be used to call the BrAPI server for
AlleleMatrix entities}

\item{\code{breedingMethods}}{Get the BreedingMethods R6 class object which can be used to call the BrAPI server for
BreedingMethod entities}

\item{\code{calls}}{Get the Calls R6 class object which can be used to call the BrAPI server for
Call entities}

\item{\code{callSets}}{Get the CallSets R6 class object which can be used to call the BrAPI server for
CallSet entities}

\item{\code{crosses}}{Get the Crosses R6 class object which can be used to call the BrAPI server for
Cross entities}

\item{\code{crossingProjects}}{Get the CrossingProjects R6 class object which can be used to call the BrAPI server for
CrossingProject entities}

\item{\code{events}}{Get the Events R6 class object which can be used to call the BrAPI server for
Event entities}

\item{\code{genomeMaps}}{Get the GenomeMaps R6 class object which can be used to call the BrAPI server for
GenomeMap entities}

\item{\code{germplasm}}{Get the Germplasm R6 class object which can be used to call the BrAPI server for
Germplasm entities}

\item{\code{germplasmAttributes}}{Get the GermplasmAttributes R6 class object which can be used to call the BrAPI server for
GermplasmAttribute entities}

\item{\code{germplasmAttributeValues}}{Get the GermplasmAttributeValues R6 class object which can be used to call the BrAPI server for
GermplasmAttributeValue entities}

\item{\code{images}}{Get the Images R6 class object which can be used to call the BrAPI server for
Image entities}

\item{\code{lists}}{Get the Lists R6 class object which can be used to call the BrAPI server for
List entities}

\item{\code{locations}}{Get the Locations R6 class object which can be used to call the BrAPI server for
Location entities}

\item{\code{markerPositions}}{Get the MarkerPositions R6 class object which can be used to call the BrAPI server for
MarkerPosition entities}

\item{\code{methods}}{Get the Methods R6 class object which can be used to call the BrAPI server for
Method entities}

\item{\code{observations}}{Get the Observations R6 class object which can be used to call the BrAPI server for
Observation entities}

\item{\code{observationUnits}}{Get the ObservationUnits R6 class object which can be used to call the BrAPI server for
ObservationUnit entities}

\item{\code{observationVariables}}{Get the ObservationVariables R6 class object which can be used to call the BrAPI server for
ObservationVariable entities}

\item{\code{ontologies}}{Get the Ontologies R6 class object which can be used to call the BrAPI server for
Ontology entities}

\item{\code{pedigreeNodes}}{Get the PedigreeNodes R6 class object which can be used to call the BrAPI server for
PedigreeNode entities}

\item{\code{people}}{Get the People R6 class object which can be used to call the BrAPI server for
Person entities}

\item{\code{plannedCrosses}}{Get the PlannedCrosses R6 class object which can be used to call the BrAPI server for
PlannedCross entities}

\item{\code{plates}}{Get the Plates R6 class object which can be used to call the BrAPI server for
Plate entities}

\item{\code{programs}}{Get the Programs R6 class object which can be used to call the BrAPI server for
Program entities}

\item{\code{references}}{Get the References R6 class object which can be used to call the BrAPI server for
Reference entities}

\item{\code{referenceSets}}{Get the ReferenceSets R6 class object which can be used to call the BrAPI server for
ReferenceSet entities}

\item{\code{samples}}{Get the Samples R6 class object which can be used to call the BrAPI server for
Sample entities}

\item{\code{scales}}{Get the Scales R6 class object which can be used to call the BrAPI server for
Scale entities}

\item{\code{seasons}}{Get the Seasons R6 class object which can be used to call the BrAPI server for
Season entities}

\item{\code{seedLots}}{Get the SeedLots R6 class object which can be used to call the BrAPI server for
SeedLot entities}

\item{\code{studies}}{Get the Studies R6 class object which can be used to call the BrAPI server for
Study entities}

\item{\code{traits}}{Get the Traits R6 class object which can be used to call the BrAPI server for
Trait entities}

\item{\code{trials}}{Get the Trials R6 class object which can be used to call the BrAPI server for
Trial entities}

\item{\code{variants}}{Get the Variants R6 class object which can be used to call the BrAPI server for
Variant entities}

\item{\code{variantSets}}{Get the VariantSets R6 class object which can be used to call the BrAPI server for
VariantSet entities}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BrAPIClient-clone}{\code{BrAPIClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="BrAPI.R6" data-topic="BaseBrAPIClient" data-id="initialize"><a href='../../BrAPI.R6/html/BaseBrAPIClient.html#method-BaseBrAPIClient-initialize'><code>BrAPI.R6::BaseBrAPIClient$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BrAPI.R6" data-topic="BaseBrAPIClient" data-id="perform_get_request"><a href='../../BrAPI.R6/html/BaseBrAPIClient.html#method-BaseBrAPIClient-perform_get_request'><code>BrAPI.R6::BaseBrAPIClient$perform_get_request()</code></a></span></li>
<li><span class="pkg-link" data-pkg="BrAPI.R6" data-topic="BaseBrAPIClient" data-id="perform_post_request"><a href='../../BrAPI.R6/html/BaseBrAPIClient.html#method-BaseBrAPIClient-perform_post_request'><code>BrAPI.R6::BaseBrAPIClient$perform_post_request()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BrAPIClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BrAPIClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BrAPIClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
