% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_BreedingMethod.R
\docType{class}
\name{BreedingMethods}
\alias{BreedingMethods}
\title{BreedingMethods Class}
\description{
The BreedingMethods class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the BreedingMethod entity
}
\details{
The BreedingMethods class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the BreedingMethod entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BreedingMethods-new}{\code{BreedingMethods$new()}}
\item \href{#method-BreedingMethods-get}{\code{BreedingMethods$get()}}
\item \href{#method-BreedingMethods-getAll}{\code{BreedingMethods$getAll()}}
\item \href{#method-BreedingMethods-clone}{\code{BreedingMethods$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BreedingMethods-new"></a>}}
\if{latex}{\out{\hypertarget{method-BreedingMethods-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BreedingMethods$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BreedingMethods-get"></a>}}
\if{latex}{\out{\hypertarget{method-BreedingMethods-get}{}}}
\subsection{Method \code{get()}}{
Gets a BreedingMethod object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BreedingMethods$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the BreedingMethod to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a BreedingMethod object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BreedingMethods-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-BreedingMethods-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list BreedingMethod objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BreedingMethods$getAll(page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of BreedingMethod objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BreedingMethods-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BreedingMethods-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BreedingMethods$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
