% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Call.R
\docType{class}
\name{Calls}
\alias{Calls}
\title{Calls Class}
\description{
The Calls class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Call entity
}
\details{
The Calls class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Call entity
}
\examples{
Calls$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Calls-new}{\code{Calls$new()}}
\item \href{#method-Calls-getAll}{\code{Calls$getAll()}}
\item \href{#method-Calls-search}{\code{Calls$search()}}
\item \href{#method-Calls-searchResult}{\code{Calls$searchResult()}}
\item \href{#method-Calls-clone}{\code{Calls$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calls-new"></a>}}
\if{latex}{\out{\hypertarget{method-Calls-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calls$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calls-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Calls-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Call objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calls$getAll(
  callSetDbIds = NULL,
  variantDbIds = NULL,
  variantSetDbIds = NULL,
  expandHomozygotes = NULL,
  sepPhased = NULL,
  sepUnphased = NULL,
  unknownString = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{callSetDbIds}}{A list of IDs which uniquely identify \code{CallSets} within the given database server}

\item{\code{variantDbIds}}{A list of IDs which uniquely identify \code{Variant} within the given database server}

\item{\code{variantSetDbIds}}{A list of IDs which uniquely identify \code{VariantSets} within the given database server}

\item{\code{expandHomozygotes}}{Should homozygotes be expanded (true) or collapsed into a single occurrence (false)}

\item{\code{sepPhased}}{The string used as a separator for phased allele calls.}

\item{\code{sepUnphased}}{The string used as a separator for unphased allele calls.}

\item{\code{unknownString}}{The string used as a representation for missing data.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Call objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calls-search"></a>}}
\if{latex}{\out{\hypertarget{method-Calls-search}{}}}
\subsection{Method \code{search()}}{
Searches for Call objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calls$search(
  callSetDbIds = NULL,
  variantDbIds = NULL,
  variantSetDbIds = NULL,
  expandHomozygotes = NULL,
  sepPhased = NULL,
  sepUnphased = NULL,
  unknownString = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{callSetDbIds}}{A list of IDs which uniquely identify \code{CallSets} within the given database server}

\item{\code{variantDbIds}}{A list of IDs which uniquely identify \code{Variant} within the given database server}

\item{\code{variantSetDbIds}}{A list of IDs which uniquely identify \code{VariantSets} within the given database server}

\item{\code{expandHomozygotes}}{Should homozygotes be expanded (true) or collapsed into a single occurrence (false)}

\item{\code{sepPhased}}{The string used as a separator for phased allele calls.}

\item{\code{sepUnphased}}{The string used as a separator for unphased allele calls.}

\item{\code{unknownString}}{The string used as a representation for missing data.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Call objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calls-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Calls-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calls$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calls-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Calls-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calls$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
