% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_CallSet.R
\docType{class}
\name{CallSets}
\alias{CallSets}
\title{CallSets Class}
\description{
The CallSets class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the CallSet entity
}
\details{
The CallSets class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the CallSet entity
}
\examples{
CallSets$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CallSets-new}{\code{CallSets$new()}}
\item \href{#method-CallSets-get}{\code{CallSets$get()}}
\item \href{#method-CallSets-getAll}{\code{CallSets$getAll()}}
\item \href{#method-CallSets-search}{\code{CallSets$search()}}
\item \href{#method-CallSets-searchResult}{\code{CallSets$searchResult()}}
\item \href{#method-CallSets-clone}{\code{CallSets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallSets-new"></a>}}
\if{latex}{\out{\hypertarget{method-CallSets-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallSets$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallSets-get"></a>}}
\if{latex}{\out{\hypertarget{method-CallSets-get}{}}}
\subsection{Method \code{get()}}{
Gets a CallSet object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallSets$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the CallSet to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a CallSet object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallSets-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-CallSets-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list CallSet objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallSets$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  sampleDbIds = NULL,
  sampleNames = NULL,
  callSetDbIds = NULL,
  callSetNames = NULL,
  variantSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{sampleDbIds}}{A list of IDs which uniquely identify \code{Samples} within the given database server}

\item{\code{sampleNames}}{A list of human readable names associated with \code{Samples}}

\item{\code{callSetDbIds}}{A list of IDs which uniquely identify \code{CallSets} within the given database server}

\item{\code{callSetNames}}{A list of human readable names associated with \code{CallSets}}

\item{\code{variantSetDbIds}}{A list of IDs which uniquely identify \code{VariantSets} within the given database server}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of CallSet objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallSets-search"></a>}}
\if{latex}{\out{\hypertarget{method-CallSets-search}{}}}
\subsection{Method \code{search()}}{
Searches for CallSet objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallSets$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  sampleDbIds = NULL,
  sampleNames = NULL,
  callSetDbIds = NULL,
  callSetNames = NULL,
  variantSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{sampleDbIds}}{A list of IDs which uniquely identify \code{Samples} within the given database server}

\item{\code{sampleNames}}{A list of human readable names associated with \code{Samples}}

\item{\code{callSetDbIds}}{A list of IDs which uniquely identify \code{CallSets} within the given database server}

\item{\code{callSetNames}}{A list of human readable names associated with \code{CallSets}}

\item{\code{variantSetDbIds}}{A list of IDs which uniquely identify \code{VariantSets} within the given database server
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for CallSet objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallSets-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-CallSets-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallSets$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallSets-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CallSets-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallSets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
