% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{client_credentials}
\alias{client_credentials}
\title{Creates an OAuth Client Credentials provider for BrAPIClient R6 Class object.}
\usage{
client_credentials(
  id,
  secret = NULL,
  token_url,
  name = NULL,
  scope = NULL,
  token_params = NULL
)
}
\arguments{
\item{id}{The client ID to be used for OAuth Authorization.}

\item{secret}{The client secret to be used for OAuth Authorization.}

\item{token_url}{The token URL to be used for OAuth Authorization.}

\item{name}{The name to be used for the OAuth client.}

\item{scope}{The scope to be used for OAuth Authorization.}

\item{token_params}{Additional token parameters to be used for OAuth Authorization.}
}
\value{
A configured OAuth provider R6 Class object.
}
\description{
Creates an OAuth Client Credentials provider for BrAPIClient R6 Class object.
}
\examples{
\dontrun{
oauth_flow("client_id", "client_secret", "https://example.com/token")
}
}
\seealso{
Other authentication: 
\code{\link{basic_authentication}()},
\code{\link{oauth_flow}()}
}
\concept{authentication}
\keyword{authentication}
