% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Cross.R
\docType{class}
\name{Crosses}
\alias{Crosses}
\title{Crosses Class}
\description{
The Crosses class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Cross entity
}
\details{
The Crosses class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Cross entity
}
\examples{
Crosses$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Crosses-new}{\code{Crosses$new()}}
\item \href{#method-Crosses-getAll}{\code{Crosses$getAll()}}
\item \href{#method-Crosses-create}{\code{Crosses$create()}}
\item \href{#method-Crosses-update}{\code{Crosses$update()}}
\item \href{#method-Crosses-clone}{\code{Crosses$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crosses-new"></a>}}
\if{latex}{\out{\hypertarget{method-Crosses-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosses$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crosses-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Crosses-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Cross objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosses$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  crossingProjectDbIds = NULL,
  crossingProjectNames = NULL,
  crossDbIds = NULL,
  crossNames = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{crossingProjectDbIds}}{Search for Crossing Projects with this unique id}

\item{\code{crossingProjectNames}}{The human readable name for a crossing project}

\item{\code{crossDbIds}}{Search for Cross with this unique id}

\item{\code{crossNames}}{Search for Cross with this human readable name}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Cross objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crosses-create"></a>}}
\if{latex}{\out{\hypertarget{method-Crosses-create}{}}}
\subsection{Method \code{create()}}{
Creates a Cross object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosses$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Cross object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crosses-update"></a>}}
\if{latex}{\out{\hypertarget{method-Crosses-update}{}}}
\subsection{Method \code{update()}}{
Creates a Cross object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosses$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Cross object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Crosses-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Crosses-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Crosses$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
