% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_CrossingProject.R
\docType{class}
\name{CrossingProjects}
\alias{CrossingProjects}
\title{CrossingProjects Class}
\description{
The CrossingProjects class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the CrossingProject entity
}
\details{
The CrossingProjects class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the CrossingProject entity
}
\examples{
CrossingProjects$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CrossingProjects-new}{\code{CrossingProjects$new()}}
\item \href{#method-CrossingProjects-get}{\code{CrossingProjects$get()}}
\item \href{#method-CrossingProjects-getAll}{\code{CrossingProjects$getAll()}}
\item \href{#method-CrossingProjects-create}{\code{CrossingProjects$create()}}
\item \href{#method-CrossingProjects-update}{\code{CrossingProjects$update()}}
\item \href{#method-CrossingProjects-clone}{\code{CrossingProjects$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossingProjects-new"></a>}}
\if{latex}{\out{\hypertarget{method-CrossingProjects-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossingProjects$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossingProjects-get"></a>}}
\if{latex}{\out{\hypertarget{method-CrossingProjects-get}{}}}
\subsection{Method \code{get()}}{
Gets a CrossingProject object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossingProjects$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the CrossingProject to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a CrossingProject object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossingProjects-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-CrossingProjects-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list CrossingProject objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossingProjects$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  crossingProjectDbIds = NULL,
  crossingProjectNames = NULL,
  includePotentialParents = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{crossingProjectDbIds}}{Search for Crossing Projects with this unique id}

\item{\code{crossingProjectNames}}{The human readable name for a crossing project}

\item{\code{includePotentialParents}}{If the parameter 'includePotentialParents' is false, the array 'potentialParents' should be empty, null, or excluded from the response object.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of CrossingProject objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossingProjects-create"></a>}}
\if{latex}{\out{\hypertarget{method-CrossingProjects-create}{}}}
\subsection{Method \code{create()}}{
Creates a CrossingProject object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossingProjects$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new CrossingProject object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossingProjects-update"></a>}}
\if{latex}{\out{\hypertarget{method-CrossingProjects-update}{}}}
\subsection{Method \code{update()}}{
Creates a CrossingProject object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossingProjects$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated CrossingProject object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CrossingProjects-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CrossingProjects-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CrossingProjects$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
