% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_GenomeMap.R
\docType{class}
\name{GenomeMaps}
\alias{GenomeMaps}
\title{GenomeMaps Class}
\description{
The GenomeMaps class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the GenomeMap entity
}
\details{
The GenomeMaps class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the GenomeMap entity
}
\examples{
GenomeMaps$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenomeMaps-new}{\code{GenomeMaps$new()}}
\item \href{#method-GenomeMaps-get}{\code{GenomeMaps$get()}}
\item \href{#method-GenomeMaps-getAll}{\code{GenomeMaps$getAll()}}
\item \href{#method-GenomeMaps-clone}{\code{GenomeMaps$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeMaps-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeMaps-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeMaps$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeMaps-get"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeMaps-get}{}}}
\subsection{Method \code{get()}}{
Gets a GenomeMap object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeMaps$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the GenomeMap to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a GenomeMap object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeMaps-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeMaps-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list GenomeMap objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeMaps$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  mapDbIds = NULL,
  mapPUI = NULL,
  scientificName = NULL,
  types = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{mapDbIds}}{The ID which uniquely identifies a \code{GenomeMap}}

\item{\code{mapPUI}}{The DOI or other permanent identifier for a \code{GenomeMap}}

\item{\code{scientificName}}{Full scientific binomial format name.}

\item{\code{types}}{The type of map, usually "Genetic" or "Physical"}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of GenomeMap objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenomeMaps-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenomeMaps-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenomeMaps$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
