% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_GermplasmAttribute.R
\docType{class}
\name{GermplasmAttributes}
\alias{GermplasmAttributes}
\title{GermplasmAttributes Class}
\description{
The GermplasmAttributes class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the GermplasmAttribute entity
}
\details{
The GermplasmAttributes class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the GermplasmAttribute entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GermplasmAttributes-new}{\code{GermplasmAttributes$new()}}
\item \href{#method-GermplasmAttributes-get}{\code{GermplasmAttributes$get()}}
\item \href{#method-GermplasmAttributes-getAll}{\code{GermplasmAttributes$getAll()}}
\item \href{#method-GermplasmAttributes-search}{\code{GermplasmAttributes$search()}}
\item \href{#method-GermplasmAttributes-searchResult}{\code{GermplasmAttributes$searchResult()}}
\item \href{#method-GermplasmAttributes-create}{\code{GermplasmAttributes$create()}}
\item \href{#method-GermplasmAttributes-update}{\code{GermplasmAttributes$update()}}
\item \href{#method-GermplasmAttributes-clone}{\code{GermplasmAttributes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-new"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-get"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-get}{}}}
\subsection{Method \code{get()}}{
Gets a GermplasmAttribute object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the GermplasmAttribute to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a GermplasmAttribute object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list GermplasmAttribute objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$getAll(
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  studyDbId = NULL,
  ontologyDbIds = NULL,
  methodDbIds = NULL,
  methodNames = NULL,
  methodPUIs = NULL,
  scaleDbIds = NULL,
  scaleNames = NULL,
  scalePUIs = NULL,
  dataTypes = NULL,
  traitClasses = NULL,
  traitDbIds = NULL,
  traitNames = NULL,
  traitPUIs = NULL,
  traitAttributes = NULL,
  traitAttributePUIs = NULL,
  traitEntities = NULL,
  traitEntityPUIs = NULL,
  attributeDbIds = NULL,
  attributeNames = NULL,
  attributePUIs = NULL,
  attributeCategories = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{studyDbId}}{**Deprecated in v2.}

\item{\code{ontologyDbIds}}{List of ontology IDs to search for}

\item{\code{methodDbIds}}{List of methods to filter search results}

\item{\code{methodNames}}{Human readable name for the method <br/>MIAPPE V1.}

\item{\code{methodPUIs}}{The Permanent Unique Identifier of a Method, usually in the form of a URI}

\item{\code{scaleDbIds}}{The unique identifier for a Scale}

\item{\code{scaleNames}}{Name of the scale <br/>MIAPPE V1.}

\item{\code{scalePUIs}}{The Permanent Unique Identifier of a Scale, usually in the form of a URI}

\item{\code{dataTypes}}{List of scale data types to filter search results}

\item{\code{traitClasses}}{List of trait classes to filter search results}

\item{\code{traitDbIds}}{The unique identifier for a Trait}

\item{\code{traitNames}}{The human readable name of a trait <br/>MIAPPE V1.}

\item{\code{traitPUIs}}{The Permanent Unique Identifier of a Trait, usually in the form of a URI}

\item{\code{traitAttributes}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitAttributePUIs}}{The Permanent Unique Identifier of a Trait Attribute, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitEntities}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.}

\item{\code{traitEntityPUIs}}{The Permanent Unique Identifier of a Trait Entity, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.}

\item{\code{attributeDbIds}}{List of Germplasm Attribute IDs to search for}

\item{\code{attributeNames}}{List of human readable Germplasm Attribute names to search for}

\item{\code{attributePUIs}}{The Permanent Unique Identifier of an Attribute, usually in the form of a URI}

\item{\code{attributeCategories}}{General category for the attribute.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of GermplasmAttribute objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-search"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-search}{}}}
\subsection{Method \code{search()}}{
Searches for GermplasmAttribute objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$search(
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  studyDbId = NULL,
  ontologyDbIds = NULL,
  methodDbIds = NULL,
  methodNames = NULL,
  methodPUIs = NULL,
  scaleDbIds = NULL,
  scaleNames = NULL,
  scalePUIs = NULL,
  dataTypes = NULL,
  traitClasses = NULL,
  traitDbIds = NULL,
  traitNames = NULL,
  traitPUIs = NULL,
  traitAttributes = NULL,
  traitAttributePUIs = NULL,
  traitEntities = NULL,
  traitEntityPUIs = NULL,
  attributeDbIds = NULL,
  attributeNames = NULL,
  attributePUIs = NULL,
  attributeCategories = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{studyDbId}}{**Deprecated in v2.}

\item{\code{ontologyDbIds}}{List of ontology IDs to search for}

\item{\code{methodDbIds}}{List of methods to filter search results}

\item{\code{methodNames}}{Human readable name for the method <br/>MIAPPE V1.}

\item{\code{methodPUIs}}{The Permanent Unique Identifier of a Method, usually in the form of a URI}

\item{\code{scaleDbIds}}{The unique identifier for a Scale}

\item{\code{scaleNames}}{Name of the scale <br/>MIAPPE V1.}

\item{\code{scalePUIs}}{The Permanent Unique Identifier of a Scale, usually in the form of a URI}

\item{\code{dataTypes}}{List of scale data types to filter search results}

\item{\code{traitClasses}}{List of trait classes to filter search results}

\item{\code{traitDbIds}}{The unique identifier for a Trait}

\item{\code{traitNames}}{The human readable name of a trait <br/>MIAPPE V1.}

\item{\code{traitPUIs}}{The Permanent Unique Identifier of a Trait, usually in the form of a URI}

\item{\code{traitAttributes}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitAttributePUIs}}{The Permanent Unique Identifier of a Trait Attribute, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitEntities}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.}

\item{\code{traitEntityPUIs}}{The Permanent Unique Identifier of a Trait Entity, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.}

\item{\code{attributeDbIds}}{List of Germplasm Attribute IDs to search for}

\item{\code{attributeNames}}{List of human readable Germplasm Attribute names to search for}

\item{\code{attributePUIs}}{The Permanent Unique Identifier of an Attribute, usually in the form of a URI}

\item{\code{attributeCategories}}{General category for the attribute.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for GermplasmAttribute objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-create"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-create}{}}}
\subsection{Method \code{create()}}{
Creates a GermplasmAttribute object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new GermplasmAttribute object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-update"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-update}{}}}
\subsection{Method \code{update()}}{
Creates a GermplasmAttribute object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated GermplasmAttribute object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
