% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_GermplasmAttributeValue.R
\docType{class}
\name{GermplasmAttributeValues}
\alias{GermplasmAttributeValues}
\title{GermplasmAttributeValues Class}
\description{
The GermplasmAttributeValues class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the GermplasmAttributeValue entity
}
\details{
The GermplasmAttributeValues class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the GermplasmAttributeValue entity
}
\examples{
GermplasmAttributeValues$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GermplasmAttributeValues-new}{\code{GermplasmAttributeValues$new()}}
\item \href{#method-GermplasmAttributeValues-get}{\code{GermplasmAttributeValues$get()}}
\item \href{#method-GermplasmAttributeValues-getAll}{\code{GermplasmAttributeValues$getAll()}}
\item \href{#method-GermplasmAttributeValues-search}{\code{GermplasmAttributeValues$search()}}
\item \href{#method-GermplasmAttributeValues-searchResult}{\code{GermplasmAttributeValues$searchResult()}}
\item \href{#method-GermplasmAttributeValues-create}{\code{GermplasmAttributeValues$create()}}
\item \href{#method-GermplasmAttributeValues-update}{\code{GermplasmAttributeValues$update()}}
\item \href{#method-GermplasmAttributeValues-clone}{\code{GermplasmAttributeValues$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-new"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-get"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-get}{}}}
\subsection{Method \code{get()}}{
Gets a GermplasmAttributeValue object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the GermplasmAttributeValue to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a GermplasmAttributeValue object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list GermplasmAttributeValue objects from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  attributeValueDbIds = NULL,
  attributeDbIds = NULL,
  attributeNames = NULL,
  ontologyDbIds = NULL,
  methodDbIds = NULL,
  scaleDbIds = NULL,
  traitDbIds = NULL,
  traitClasses = NULL,
  dataTypes = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{attributeValueDbIds}}{List of Germplasm Attribute Value IDs to search for}

\item{\code{attributeDbIds}}{List of Germplasm Attribute IDs to search for}

\item{\code{attributeNames}}{List of human readable Germplasm Attribute names to search for}

\item{\code{ontologyDbIds}}{List of ontology IDs to search for}

\item{\code{methodDbIds}}{List of methods to filter search results}

\item{\code{scaleDbIds}}{List of scales to filter search results}

\item{\code{traitDbIds}}{List of trait unique ID to filter search results}

\item{\code{traitClasses}}{List of trait classes to filter search results}

\item{\code{dataTypes}}{List of scale data types to filter search results}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of GermplasmAttributeValue objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-search"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-search}{}}}
\subsection{Method \code{search()}}{
Searches for GermplasmAttributeValue objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  attributeValueDbIds = NULL,
  attributeDbIds = NULL,
  attributeNames = NULL,
  ontologyDbIds = NULL,
  methodDbIds = NULL,
  scaleDbIds = NULL,
  traitDbIds = NULL,
  traitClasses = NULL,
  dataTypes = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{attributeValueDbIds}}{List of Germplasm Attribute Value IDs to search for}

\item{\code{attributeDbIds}}{List of Germplasm Attribute IDs to search for}

\item{\code{attributeNames}}{List of human readable Germplasm Attribute names to search for}

\item{\code{ontologyDbIds}}{List of ontology IDs to search for}

\item{\code{methodDbIds}}{List of methods to filter search results}

\item{\code{scaleDbIds}}{List of scales to filter search results}

\item{\code{traitDbIds}}{List of trait unique ID to filter search results}

\item{\code{traitClasses}}{List of trait classes to filter search results}

\item{\code{dataTypes}}{List of scale data types to filter search results
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for GermplasmAttributeValue objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$searchResult(
  searchResultId,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-create"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-create}{}}}
\subsection{Method \code{create()}}{
Creates a GermplasmAttributeValue object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new GermplasmAttributeValue object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-update"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-update}{}}}
\subsection{Method \code{update()}}{
Creates a GermplasmAttributeValue object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated GermplasmAttributeValue object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GermplasmAttributeValues-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GermplasmAttributeValues-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GermplasmAttributeValues$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
