% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{guess_username}
\alias{guess_username}
\title{Attempts to guess the username from the system environment}
\usage{
guess_username()
}
\value{
The guessed username
}
\description{
Attempts to guess the username from the system environment
}
\examples{
\dontrun{
guess_username()
}
}
\concept{utilities}
\keyword{utility}
