% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Image.R
\docType{class}
\name{Images}
\alias{Images}
\title{Images Class}
\description{
The Images class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Image entity
}
\details{
The Images class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Image entity
}
\examples{
Images$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Images-new}{\code{Images$new()}}
\item \href{#method-Images-get}{\code{Images$get()}}
\item \href{#method-Images-getAll}{\code{Images$getAll()}}
\item \href{#method-Images-search}{\code{Images$search()}}
\item \href{#method-Images-searchResult}{\code{Images$searchResult()}}
\item \href{#method-Images-create}{\code{Images$create()}}
\item \href{#method-Images-update}{\code{Images$update()}}
\item \href{#method-Images-clone}{\code{Images$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-new"></a>}}
\if{latex}{\out{\hypertarget{method-Images-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-get"></a>}}
\if{latex}{\out{\hypertarget{method-Images-get}{}}}
\subsection{Method \code{get()}}{
Gets a Image object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Image to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Image object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Images-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Image objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  descriptiveOntologyTerms = NULL,
  imageFileNames = NULL,
  imageFileSizeMax = NULL,
  imageFileSizeMin = NULL,
  imageHeightMax = NULL,
  imageHeightMin = NULL,
  imageLocation = NULL,
  imageNames = NULL,
  imageTimeStampRangeEnd = NULL,
  imageTimeStampRangeStart = NULL,
  imageWidthMax = NULL,
  imageWidthMin = NULL,
  mimeTypes = NULL,
  observationDbIds = NULL,
  imageDbIds = NULL,
  observationUnitDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{descriptiveOntologyTerms}}{A list of terms to formally describe the image to search for.}

\item{\code{imageFileNames}}{Image file names to search for.}

\item{\code{imageFileSizeMax}}{A maximum image file size to search for.}

\item{\code{imageFileSizeMin}}{A minimum image file size to search for.}

\item{\code{imageHeightMax}}{A maximum image height to search for.}

\item{\code{imageHeightMin}}{A minimum image height to search for.}

\item{\code{imageLocation}}{}

\item{\code{imageNames}}{Human readable names to search for.}

\item{\code{imageTimeStampRangeEnd}}{The latest timestamp to search for.}

\item{\code{imageTimeStampRangeStart}}{The earliest timestamp to search for.}

\item{\code{imageWidthMax}}{A maximum image width to search for.}

\item{\code{imageWidthMin}}{A minimum image width to search for.}

\item{\code{mimeTypes}}{A set of image file types to search for.}

\item{\code{observationDbIds}}{A list of observation Ids this image is associated with to search for}

\item{\code{imageDbIds}}{A list of image Ids to search for}

\item{\code{observationUnitDbIds}}{A set of observation unit identifiers to search for.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Image objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-search"></a>}}
\if{latex}{\out{\hypertarget{method-Images-search}{}}}
\subsection{Method \code{search()}}{
Searches for Image objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$search(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  descriptiveOntologyTerms = NULL,
  imageFileNames = NULL,
  imageFileSizeMax = NULL,
  imageFileSizeMin = NULL,
  imageHeightMax = NULL,
  imageHeightMin = NULL,
  imageLocation = NULL,
  imageNames = NULL,
  imageTimeStampRangeEnd = NULL,
  imageTimeStampRangeStart = NULL,
  imageWidthMax = NULL,
  imageWidthMin = NULL,
  mimeTypes = NULL,
  observationDbIds = NULL,
  imageDbIds = NULL,
  observationUnitDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{descriptiveOntologyTerms}}{A list of terms to formally describe the image to search for.}

\item{\code{imageFileNames}}{Image file names to search for.}

\item{\code{imageFileSizeMax}}{A maximum image file size to search for.}

\item{\code{imageFileSizeMin}}{A minimum image file size to search for.}

\item{\code{imageHeightMax}}{A maximum image height to search for.}

\item{\code{imageHeightMin}}{A minimum image height to search for.}

\item{\code{imageLocation}}{}

\item{\code{imageNames}}{Human readable names to search for.}

\item{\code{imageTimeStampRangeEnd}}{The latest timestamp to search for.}

\item{\code{imageTimeStampRangeStart}}{The earliest timestamp to search for.}

\item{\code{imageWidthMax}}{A maximum image width to search for.}

\item{\code{imageWidthMin}}{A minimum image width to search for.}

\item{\code{mimeTypes}}{A set of image file types to search for.}

\item{\code{observationDbIds}}{A list of observation Ids this image is associated with to search for}

\item{\code{imageDbIds}}{A list of image Ids to search for}

\item{\code{observationUnitDbIds}}{A set of observation unit identifiers to search for.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Image objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Images-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-create"></a>}}
\if{latex}{\out{\hypertarget{method-Images-create}{}}}
\subsection{Method \code{create()}}{
Creates a Image object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Image object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-update"></a>}}
\if{latex}{\out{\hypertarget{method-Images-update}{}}}
\subsection{Method \code{update()}}{
Creates a Image object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Image object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Images-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Images-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Images$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
