% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_List.R
\docType{class}
\name{Lists}
\alias{Lists}
\title{Lists Class}
\description{
The Lists class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the List entity
}
\details{
The Lists class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the List entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lists-new}{\code{Lists$new()}}
\item \href{#method-Lists-get}{\code{Lists$get()}}
\item \href{#method-Lists-getAll}{\code{Lists$getAll()}}
\item \href{#method-Lists-search}{\code{Lists$search()}}
\item \href{#method-Lists-searchResult}{\code{Lists$searchResult()}}
\item \href{#method-Lists-create}{\code{Lists$create()}}
\item \href{#method-Lists-update}{\code{Lists$update()}}
\item \href{#method-Lists-clone}{\code{Lists$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-get"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-get}{}}}
\subsection{Method \code{get()}}{
Gets a List object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the List to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a List object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list List objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$getAll(
  dateCreatedRangeStart = NULL,
  dateCreatedRangeEnd = NULL,
  dateModifiedRangeStart = NULL,
  dateModifiedRangeEnd = NULL,
  listDbIds = NULL,
  listNames = NULL,
  listOwnerNames = NULL,
  listOwnerPersonDbIds = NULL,
  listSources = NULL,
  listType = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dateCreatedRangeStart}}{Define the beginning for an interval of time and only include Lists that are created within this interval.}

\item{\code{dateCreatedRangeEnd}}{Define the end for an interval of time and only include Lists that are created within this interval.}

\item{\code{dateModifiedRangeStart}}{Define the beginning for an interval of time and only include Lists that are modified within this interval.}

\item{\code{dateModifiedRangeEnd}}{Define the end for an interval of time and only include Lists that are modified within this interval.}

\item{\code{listDbIds}}{An array of primary database identifiers to identify a set of Lists}

\item{\code{listNames}}{An array of human readable names to identify a set of Lists}

\item{\code{listOwnerNames}}{An array of names for the people or entities who are responsible for a set of Lists}

\item{\code{listOwnerPersonDbIds}}{An array of primary database identifiers to identify people or entities who are responsible for a set of Lists}

\item{\code{listSources}}{An array of terms identifying lists from different sources (ie 'USER', 'SYSTEM', etc)}

\item{\code{listType}}{}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of List objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-search"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-search}{}}}
\subsection{Method \code{search()}}{
Searches for List objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$search(
  dateCreatedRangeStart = NULL,
  dateCreatedRangeEnd = NULL,
  dateModifiedRangeStart = NULL,
  dateModifiedRangeEnd = NULL,
  listDbIds = NULL,
  listNames = NULL,
  listOwnerNames = NULL,
  listOwnerPersonDbIds = NULL,
  listSources = NULL,
  listType = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dateCreatedRangeStart}}{Define the beginning for an interval of time and only include Lists that are created within this interval.}

\item{\code{dateCreatedRangeEnd}}{Define the end for an interval of time and only include Lists that are created within this interval.}

\item{\code{dateModifiedRangeStart}}{Define the beginning for an interval of time and only include Lists that are modified within this interval.}

\item{\code{dateModifiedRangeEnd}}{Define the end for an interval of time and only include Lists that are modified within this interval.}

\item{\code{listDbIds}}{An array of primary database identifiers to identify a set of Lists}

\item{\code{listNames}}{An array of human readable names to identify a set of Lists}

\item{\code{listOwnerNames}}{An array of names for the people or entities who are responsible for a set of Lists}

\item{\code{listOwnerPersonDbIds}}{An array of primary database identifiers to identify people or entities who are responsible for a set of Lists}

\item{\code{listSources}}{An array of terms identifying lists from different sources (ie 'USER', 'SYSTEM', etc)}

\item{\code{listType}}{using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for List objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-create"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-create}{}}}
\subsection{Method \code{create()}}{
Creates a List object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new List object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-update"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-update}{}}}
\subsection{Method \code{update()}}{
Creates a List object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated List object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lists-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lists-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lists$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
