% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_MarkerPosition.R
\docType{class}
\name{MarkerPositions}
\alias{MarkerPositions}
\title{MarkerPositions Class}
\description{
The MarkerPositions class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the MarkerPosition entity
}
\details{
The MarkerPositions class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the MarkerPosition entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MarkerPositions-new}{\code{MarkerPositions$new()}}
\item \href{#method-MarkerPositions-getAll}{\code{MarkerPositions$getAll()}}
\item \href{#method-MarkerPositions-search}{\code{MarkerPositions$search()}}
\item \href{#method-MarkerPositions-searchResult}{\code{MarkerPositions$searchResult()}}
\item \href{#method-MarkerPositions-clone}{\code{MarkerPositions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerPositions-new"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerPositions-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerPositions$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerPositions-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerPositions-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list MarkerPosition objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerPositions$getAll(
  mapDbIds = NULL,
  linkageGroupNames = NULL,
  variantDbIds = NULL,
  minPosition = NULL,
  maxPosition = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapDbIds}}{A list of IDs which uniquely identify \code{GenomeMaps} within the given database server}

\item{\code{linkageGroupNames}}{A list of Uniquely Identifiable linkage group names}

\item{\code{variantDbIds}}{A list of IDs which uniquely identify \code{Variants} within the given database server}

\item{\code{minPosition}}{The minimum position of markers in a given map}

\item{\code{maxPosition}}{The maximum position of markers in a given map}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of MarkerPosition objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerPositions-search"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerPositions-search}{}}}
\subsection{Method \code{search()}}{
Searches for MarkerPosition objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerPositions$search(
  mapDbIds = NULL,
  linkageGroupNames = NULL,
  variantDbIds = NULL,
  minPosition = NULL,
  maxPosition = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mapDbIds}}{A list of IDs which uniquely identify \code{GenomeMaps} within the given database server}

\item{\code{linkageGroupNames}}{A list of Uniquely Identifiable linkage group names}

\item{\code{variantDbIds}}{A list of IDs which uniquely identify \code{Variants} within the given database server}

\item{\code{minPosition}}{The minimum position of markers in a given map}

\item{\code{maxPosition}}{The maximum position of markers in a given map
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for MarkerPosition objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerPositions-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerPositions-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerPositions$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MarkerPositions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MarkerPositions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MarkerPositions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
