% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Method.R
\docType{class}
\name{Methods}
\alias{Methods}
\title{Methods Class}
\description{
The Methods class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Method entity
}
\details{
The Methods class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Method entity
}
\examples{
Methods$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Methods-new}{\code{Methods$new()}}
\item \href{#method-Methods-get}{\code{Methods$get()}}
\item \href{#method-Methods-getAll}{\code{Methods$getAll()}}
\item \href{#method-Methods-create}{\code{Methods$create()}}
\item \href{#method-Methods-update}{\code{Methods$update()}}
\item \href{#method-Methods-clone}{\code{Methods$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Methods-new"></a>}}
\if{latex}{\out{\hypertarget{method-Methods-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methods$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Methods-get"></a>}}
\if{latex}{\out{\hypertarget{method-Methods-get}{}}}
\subsection{Method \code{get()}}{
Gets a Method object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methods$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Method to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Method object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Methods-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Methods-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Method objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methods$getAll(
  commonCropNames = NULL,
  ontologyDbIds = NULL,
  programDbIds = NULL,
  programNames = NULL,
  scaleDbIds = NULL,
  observationVariableDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{ontologyDbIds}}{The unique identifier for an ontology definition.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{scaleDbIds}}{The unique identifier for a method.}

\item{\code{observationVariableDbIds}}{The unique identifier for an observation variable.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Method objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Methods-create"></a>}}
\if{latex}{\out{\hypertarget{method-Methods-create}{}}}
\subsection{Method \code{create()}}{
Creates a Method object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methods$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Method object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Methods-update"></a>}}
\if{latex}{\out{\hypertarget{method-Methods-update}{}}}
\subsection{Method \code{update()}}{
Creates a Method object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methods$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Method object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Methods-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Methods-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Methods$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
