% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\docType{class}
\name{OAuthClientCredentials}
\alias{OAuthClientCredentials}
\title{OAuthClientCredentials Class}
\description{
The OAuthClientCredentials class handles OAuth Client Credentials for httr2 requests
}
\details{
The OAuthClientCredentials class provides OAuth Client Credentials functionality for httr2 requests
}
\section{Super class}{
\code{\link[BrAPI.R6:BaseAuth]{BrAPI.R6::BaseAuth}} -> \code{OAuthClientCredentials}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OAuthClientCredentials-new}{\code{OAuthClientCredentials$new()}}
\item \href{#method-OAuthClientCredentials-authentication}{\code{OAuthClientCredentials$authentication()}}
\item \href{#method-OAuthClientCredentials-token}{\code{OAuthClientCredentials$token()}}
\item \href{#method-OAuthClientCredentials-clone}{\code{OAuthClientCredentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClientCredentials-new"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClientCredentials-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the client_credentials() function
See https://httr2.r-lib.org/reference/req_oauth_auth_code.html
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClientCredentials$new(
  id,
  secret = NULL,
  token_url,
  name = "OAuthClientCredentials",
  scope = NULL,
  token_params = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The client ID to be used for OAuth Authorization.}

\item{\code{secret}}{The client secret to be used for OAuth Authorization.}

\item{\code{token_url}}{The token URL to be used for OAuth Authorization.}

\item{\code{name}}{The name to be used for the OAuth client.}

\item{\code{scope}}{The scope to be used for OAuth Authorization.}

\item{\code{token_params}}{Additional token parameters to be used for OAuth Authorization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClientCredentials-authentication"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClientCredentials-authentication}{}}}
\subsection{Method \code{authentication()}}{
Provides the authentication token for httr2 requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClientCredentials$authentication(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{The httr2 request object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The httr2 request object with the Authorization header added
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClientCredentials-token"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClientCredentials-token}{}}}
\subsection{Method \code{token()}}{
Provides the authentication function for httr2 requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClientCredentials$token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The authentication token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClientCredentials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClientCredentials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClientCredentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
