% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\docType{class}
\name{OAuthFlow}
\alias{OAuthFlow}
\title{OAuthFlow Class}
\description{
The OAuthFlow class handles OAuth for httr2 requests
}
\details{
The OAuthFlow class provides OAuth functionality for httr2 requests
}
\section{Super class}{
\code{\link[BrAPI.R6:BaseAuth]{BrAPI.R6::BaseAuth}} -> \code{OAuthFlow}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OAuthFlow-new}{\code{OAuthFlow$new()}}
\item \href{#method-OAuthFlow-authentication}{\code{OAuthFlow$authentication()}}
\item \href{#method-OAuthFlow-clone}{\code{OAuthFlow$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthFlow-new"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthFlow-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the oauth_flow() function
See https://httr2.r-lib.org/reference/req_oauth_auth_code.html
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthFlow$new(
  id,
  secret = NULL,
  token_url,
  name = "OAuthFlow",
  auth_url,
  scope = NULL,
  pkce = TRUE,
  auth_params = list(),
  token_params = list(),
  redirect_uri = oauth_redirect_uri()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The client ID to be used for OAuth Authorization.}

\item{\code{secret}}{The client secret to be used for OAuth Authorization.}

\item{\code{token_url}}{The token URL to be used for OAuth Authorization.}

\item{\code{name}}{The name to be used for the OAuth client.}

\item{\code{auth_url}}{The authorization URL to be used for OAuth Authorization.}

\item{\code{scope}}{The scope to be used for OAuth Authorization.}

\item{\code{pkce}}{Whether to use PKCE (Proof Key for Code Exchange) for OAuth Authorization. Default is TRUE.}

\item{\code{auth_params}}{Additional authorization parameters to be used for OAuth Authorization.}

\item{\code{token_params}}{Additional token parameters to be used for OAuth Authorization.}

\item{\code{redirect_uri}}{The redirect URI to be used for OAuth Authorization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthFlow-authentication"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthFlow-authentication}{}}}
\subsection{Method \code{authentication()}}{
Provides the authentication function for httr2 requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthFlow$authentication(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{The httr2 request object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The httr2 request object with the Basic Authorization header added
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthFlow-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthFlow-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthFlow$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
