% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_ObservationUnit.R
\docType{class}
\name{ObservationUnits}
\alias{ObservationUnits}
\title{ObservationUnits Class}
\description{
The ObservationUnits class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the ObservationUnit entity
}
\details{
The ObservationUnits class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the ObservationUnit entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObservationUnits-new}{\code{ObservationUnits$new()}}
\item \href{#method-ObservationUnits-get}{\code{ObservationUnits$get()}}
\item \href{#method-ObservationUnits-getAll}{\code{ObservationUnits$getAll()}}
\item \href{#method-ObservationUnits-search}{\code{ObservationUnits$search()}}
\item \href{#method-ObservationUnits-searchResult}{\code{ObservationUnits$searchResult()}}
\item \href{#method-ObservationUnits-create}{\code{ObservationUnits$create()}}
\item \href{#method-ObservationUnits-update}{\code{ObservationUnits$update()}}
\item \href{#method-ObservationUnits-clone}{\code{ObservationUnits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-get"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-get}{}}}
\subsection{Method \code{get()}}{
Gets a ObservationUnit object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the ObservationUnit to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a ObservationUnit object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list ObservationUnit objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  programDbIds = NULL,
  programNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationUnitDbIds = NULL,
  observationUnitNames = NULL,
  observationLevels = NULL,
  observationLevelRelationships = NULL,
  includeObservations = NULL,
  seasonDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationUnitDbIds}}{The unique id of an observation unit}

\item{\code{observationUnitNames}}{The human readable identifier for an Observation Unit}

\item{\code{observationLevels}}{Searches for values in ObservationUnit->observationUnitPosition->observationLevel}

\item{\code{observationLevelRelationships}}{Searches for values in ObservationUnit->observationUnitPosition->observationLevelRelationships}

\item{\code{includeObservations}}{Use this parameter to include a list of observations embedded in each ObservationUnit object.}

\item{\code{seasonDbIds}}{The year or Phenotyping campaign of a multi-annual study (trees, grape, ...)}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of ObservationUnit objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-search"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-search}{}}}
\subsection{Method \code{search()}}{
Searches for ObservationUnit objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  programDbIds = NULL,
  programNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationUnitDbIds = NULL,
  observationUnitNames = NULL,
  observationLevels = NULL,
  observationLevelRelationships = NULL,
  includeObservations = NULL,
  seasonDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationUnitDbIds}}{The unique id of an observation unit}

\item{\code{observationUnitNames}}{The human readable identifier for an Observation Unit}

\item{\code{observationLevels}}{Searches for values in ObservationUnit->observationUnitPosition->observationLevel}

\item{\code{observationLevelRelationships}}{Searches for values in ObservationUnit->observationUnitPosition->observationLevelRelationships}

\item{\code{includeObservations}}{Use this parameter to include a list of observations embedded in each ObservationUnit object.}

\item{\code{seasonDbIds}}{The year or Phenotyping campaign of a multi-annual study (trees, grape, ...)
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for ObservationUnit objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-create"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-create}{}}}
\subsection{Method \code{create()}}{
Creates a ObservationUnit object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new ObservationUnit object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-update"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-update}{}}}
\subsection{Method \code{update()}}{
Creates a ObservationUnit object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated ObservationUnit object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationUnits-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationUnits-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationUnits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
