% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_ObservationVariable.R
\docType{class}
\name{ObservationVariables}
\alias{ObservationVariables}
\title{ObservationVariables Class}
\description{
The ObservationVariables class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the ObservationVariable entity
}
\details{
The ObservationVariables class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the ObservationVariable entity
}
\examples{
ObservationVariables$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObservationVariables-new}{\code{ObservationVariables$new()}}
\item \href{#method-ObservationVariables-get}{\code{ObservationVariables$get()}}
\item \href{#method-ObservationVariables-getAll}{\code{ObservationVariables$getAll()}}
\item \href{#method-ObservationVariables-search}{\code{ObservationVariables$search()}}
\item \href{#method-ObservationVariables-searchResult}{\code{ObservationVariables$searchResult()}}
\item \href{#method-ObservationVariables-create}{\code{ObservationVariables$create()}}
\item \href{#method-ObservationVariables-update}{\code{ObservationVariables$update()}}
\item \href{#method-ObservationVariables-clone}{\code{ObservationVariables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-get"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-get}{}}}
\subsection{Method \code{get()}}{
Gets a ObservationVariable object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the ObservationVariable to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a ObservationVariable object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list ObservationVariable objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$getAll(
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  studyDbId = NULL,
  ontologyDbIds = NULL,
  methodDbIds = NULL,
  methodNames = NULL,
  methodPUIs = NULL,
  scaleDbIds = NULL,
  scaleNames = NULL,
  scalePUIs = NULL,
  dataTypes = NULL,
  traitClasses = NULL,
  traitDbIds = NULL,
  traitNames = NULL,
  traitPUIs = NULL,
  traitAttributes = NULL,
  traitAttributePUIs = NULL,
  traitEntities = NULL,
  traitEntityPUIs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{studyDbId}}{**Deprecated in v2.}

\item{\code{ontologyDbIds}}{List of ontology IDs to search for}

\item{\code{methodDbIds}}{List of methods to filter search results}

\item{\code{methodNames}}{Human readable name for the method <br/>MIAPPE V1.}

\item{\code{methodPUIs}}{The Permanent Unique Identifier of a Method, usually in the form of a URI}

\item{\code{scaleDbIds}}{The unique identifier for a Scale}

\item{\code{scaleNames}}{Name of the scale <br/>MIAPPE V1.}

\item{\code{scalePUIs}}{The Permanent Unique Identifier of a Scale, usually in the form of a URI}

\item{\code{dataTypes}}{List of scale data types to filter search results}

\item{\code{traitClasses}}{List of trait classes to filter search results}

\item{\code{traitDbIds}}{The unique identifier for a Trait}

\item{\code{traitNames}}{The human readable name of a trait <br/>MIAPPE V1.}

\item{\code{traitPUIs}}{The Permanent Unique Identifier of a Trait, usually in the form of a URI}

\item{\code{traitAttributes}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitAttributePUIs}}{The Permanent Unique Identifier of a Trait Attribute, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitEntities}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.}

\item{\code{traitEntityPUIs}}{The Permanent Unique Identifier of a Trait Entity, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of ObservationVariable objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-search"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-search}{}}}
\subsection{Method \code{search()}}{
Searches for ObservationVariable objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$search(
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  studyDbId = NULL,
  ontologyDbIds = NULL,
  methodDbIds = NULL,
  methodNames = NULL,
  methodPUIs = NULL,
  scaleDbIds = NULL,
  scaleNames = NULL,
  scalePUIs = NULL,
  dataTypes = NULL,
  traitClasses = NULL,
  traitDbIds = NULL,
  traitNames = NULL,
  traitPUIs = NULL,
  traitAttributes = NULL,
  traitAttributePUIs = NULL,
  traitEntities = NULL,
  traitEntityPUIs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{studyDbId}}{**Deprecated in v2.}

\item{\code{ontologyDbIds}}{List of ontology IDs to search for}

\item{\code{methodDbIds}}{List of methods to filter search results}

\item{\code{methodNames}}{Human readable name for the method <br/>MIAPPE V1.}

\item{\code{methodPUIs}}{The Permanent Unique Identifier of a Method, usually in the form of a URI}

\item{\code{scaleDbIds}}{The unique identifier for a Scale}

\item{\code{scaleNames}}{Name of the scale <br/>MIAPPE V1.}

\item{\code{scalePUIs}}{The Permanent Unique Identifier of a Scale, usually in the form of a URI}

\item{\code{dataTypes}}{List of scale data types to filter search results}

\item{\code{traitClasses}}{List of trait classes to filter search results}

\item{\code{traitDbIds}}{The unique identifier for a Trait}

\item{\code{traitNames}}{The human readable name of a trait <br/>MIAPPE V1.}

\item{\code{traitPUIs}}{The Permanent Unique Identifier of a Trait, usually in the form of a URI}

\item{\code{traitAttributes}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitAttributePUIs}}{The Permanent Unique Identifier of a Trait Attribute, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the attribute is the observed feature (or characteristic) of the entity e.}

\item{\code{traitEntities}}{A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.}

\item{\code{traitEntityPUIs}}{The Permanent Unique Identifier of a Trait Entity, usually in the form of a URI <br/>A trait can be decomposed as "Trait" = "Entity" + "Attribute", the entity is the part of the plant that the trait refers to e.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for ObservationVariable objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-create"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-create}{}}}
\subsection{Method \code{create()}}{
Creates a ObservationVariable object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new ObservationVariable object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-update"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-update}{}}}
\subsection{Method \code{update()}}{
Creates a ObservationVariable object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated ObservationVariable object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationVariables-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationVariables-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationVariables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
