% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Ontology.R
\docType{class}
\name{Ontologies}
\alias{Ontologies}
\title{Ontologies Class}
\description{
The Ontologies class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Ontology entity
}
\details{
The Ontologies class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Ontology entity
}
\examples{
Ontologies$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ontologies-new}{\code{Ontologies$new()}}
\item \href{#method-Ontologies-get}{\code{Ontologies$get()}}
\item \href{#method-Ontologies-getAll}{\code{Ontologies$getAll()}}
\item \href{#method-Ontologies-create}{\code{Ontologies$create()}}
\item \href{#method-Ontologies-update}{\code{Ontologies$update()}}
\item \href{#method-Ontologies-clone}{\code{Ontologies$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontologies-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ontologies-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontologies$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontologies-get"></a>}}
\if{latex}{\out{\hypertarget{method-Ontologies-get}{}}}
\subsection{Method \code{get()}}{
Gets a Ontology object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontologies$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Ontology to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Ontology object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontologies-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Ontologies-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Ontology objects from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontologies$getAll(
  ontologyDbIds = NULL,
  ontologyNames = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ontologyDbIds}}{The unique identifier for an ontology definition.}

\item{\code{ontologyNames}}{The human readable identifier for an ontology definition.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Ontology objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontologies-create"></a>}}
\if{latex}{\out{\hypertarget{method-Ontologies-create}{}}}
\subsection{Method \code{create()}}{
Creates a Ontology object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontologies$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Ontology object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontologies-update"></a>}}
\if{latex}{\out{\hypertarget{method-Ontologies-update}{}}}
\subsection{Method \code{update()}}{
Creates a Ontology object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontologies$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Ontology object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ontologies-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Ontologies-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ontologies$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
