% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_PedigreeNode.R
\docType{class}
\name{PedigreeNodes}
\alias{PedigreeNodes}
\title{PedigreeNodes Class}
\description{
The PedigreeNodes class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the PedigreeNode entity
}
\details{
The PedigreeNodes class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the PedigreeNode entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PedigreeNodes-new}{\code{PedigreeNodes$new()}}
\item \href{#method-PedigreeNodes-getAll}{\code{PedigreeNodes$getAll()}}
\item \href{#method-PedigreeNodes-search}{\code{PedigreeNodes$search()}}
\item \href{#method-PedigreeNodes-searchResult}{\code{PedigreeNodes$searchResult()}}
\item \href{#method-PedigreeNodes-create}{\code{PedigreeNodes$create()}}
\item \href{#method-PedigreeNodes-clone}{\code{PedigreeNodes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedigreeNodes-new"></a>}}
\if{latex}{\out{\hypertarget{method-PedigreeNodes-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedigreeNodes$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedigreeNodes-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-PedigreeNodes-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list PedigreeNode objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedigreeNodes$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  germplasmPUIs = NULL,
  accessionNumbers = NULL,
  collections = NULL,
  familyCodes = NULL,
  instituteCodes = NULL,
  binomialNames = NULL,
  genus = NULL,
  species = NULL,
  synonyms = NULL,
  includeParents = NULL,
  includeSiblings = NULL,
  includeProgeny = NULL,
  includeFullTree = NULL,
  pedigreeDepth = NULL,
  progenyDepth = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{germplasmPUIs}}{List of Permanent Unique Identifiers to identify germplasm}

\item{\code{accessionNumbers}}{A collection of unique identifiers for materials or germplasm within a genebank  MCPD (v2.1) (ACCENUMB) 2.}

\item{\code{collections}}{A specific panel/collection/population name this germplasm belongs to.}

\item{\code{familyCodes}}{A familyCode representing the family this germplasm belongs to.}

\item{\code{instituteCodes}}{The code for the institute that maintains the material.}

\item{\code{binomialNames}}{List of the full binomial name (scientific name) to identify a germplasm}

\item{\code{genus}}{List of Genus names to identify germplasm}

\item{\code{species}}{List of Species names to identify germplasm}

\item{\code{synonyms}}{List of alternative names or IDs used to reference this germplasm}

\item{\code{includeParents}}{If this parameter is true, include the array of parents in the response}

\item{\code{includeSiblings}}{If this parameter is true, include the array of siblings in the response}

\item{\code{includeProgeny}}{If this parameter is true, include the array of progeny in the response}

\item{\code{includeFullTree}}{If this parameter is true, recursively include ALL of the nodes available in this pedigree tree}

\item{\code{pedigreeDepth}}{Recursively include this number of levels up the tree in the response (parents, grand-parents, great-grand-parents, etc)}

\item{\code{progenyDepth}}{Recursively include this number of levels down the tree in the response (children, grand-children, great-grand-children, etc)}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of PedigreeNode objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedigreeNodes-search"></a>}}
\if{latex}{\out{\hypertarget{method-PedigreeNodes-search}{}}}
\subsection{Method \code{search()}}{
Searches for PedigreeNode objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedigreeNodes$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  germplasmPUIs = NULL,
  accessionNumbers = NULL,
  collections = NULL,
  familyCodes = NULL,
  instituteCodes = NULL,
  binomialNames = NULL,
  genus = NULL,
  species = NULL,
  synonyms = NULL,
  includeParents = NULL,
  includeSiblings = NULL,
  includeProgeny = NULL,
  includeFullTree = NULL,
  pedigreeDepth = NULL,
  progenyDepth = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{germplasmPUIs}}{List of Permanent Unique Identifiers to identify germplasm}

\item{\code{accessionNumbers}}{A collection of unique identifiers for materials or germplasm within a genebank  MCPD (v2.1) (ACCENUMB) 2.}

\item{\code{collections}}{A specific panel/collection/population name this germplasm belongs to.}

\item{\code{familyCodes}}{A familyCode representing the family this germplasm belongs to.}

\item{\code{instituteCodes}}{The code for the institute that maintains the material.}

\item{\code{binomialNames}}{List of the full binomial name (scientific name) to identify a germplasm}

\item{\code{genus}}{List of Genus names to identify germplasm}

\item{\code{species}}{List of Species names to identify germplasm}

\item{\code{synonyms}}{List of alternative names or IDs used to reference this germplasm}

\item{\code{includeParents}}{If this parameter is true, include the array of parents in the response}

\item{\code{includeSiblings}}{If this parameter is true, include the array of siblings in the response}

\item{\code{includeProgeny}}{If this parameter is true, include the array of progeny in the response}

\item{\code{includeFullTree}}{If this parameter is true, recursively include ALL of the nodes available in this pedigree tree}

\item{\code{pedigreeDepth}}{Recursively include this number of levels up the tree in the response (parents, grand-parents, great-grand-parents, etc)}

\item{\code{progenyDepth}}{Recursively include this number of levels down the tree in the response (children, grand-children, great-grand-children, etc)
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for PedigreeNode objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedigreeNodes-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-PedigreeNodes-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedigreeNodes$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedigreeNodes-create"></a>}}
\if{latex}{\out{\hypertarget{method-PedigreeNodes-create}{}}}
\subsection{Method \code{create()}}{
Creates a PedigreeNode object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedigreeNodes$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new PedigreeNode object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PedigreeNodes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PedigreeNodes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PedigreeNodes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
