% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Person.R
\docType{class}
\name{People}
\alias{People}
\title{People Class}
\description{
The People class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Person entity
}
\details{
The People class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Person entity
}
\examples{
People$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-People-new}{\code{People$new()}}
\item \href{#method-People-get}{\code{People$get()}}
\item \href{#method-People-getAll}{\code{People$getAll()}}
\item \href{#method-People-search}{\code{People$search()}}
\item \href{#method-People-searchResult}{\code{People$searchResult()}}
\item \href{#method-People-create}{\code{People$create()}}
\item \href{#method-People-update}{\code{People$update()}}
\item \href{#method-People-clone}{\code{People$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-new"></a>}}
\if{latex}{\out{\hypertarget{method-People-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-get"></a>}}
\if{latex}{\out{\hypertarget{method-People-get}{}}}
\subsection{Method \code{get()}}{
Gets a Person object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Person to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Person object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-People-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Person objects from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  emailAddresses = NULL,
  firstNames = NULL,
  lastNames = NULL,
  mailingAddresses = NULL,
  middleNames = NULL,
  personDbIds = NULL,
  phoneNumbers = NULL,
  userIDs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{emailAddresses}}{email address for this person}

\item{\code{firstNames}}{Persons first name}

\item{\code{lastNames}}{Persons last name}

\item{\code{mailingAddresses}}{physical address of this person}

\item{\code{middleNames}}{Persons middle name}

\item{\code{personDbIds}}{Unique ID for this person}

\item{\code{phoneNumbers}}{phone number of this person}

\item{\code{userIDs}}{A systems user ID associated with this person.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Person objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-search"></a>}}
\if{latex}{\out{\hypertarget{method-People-search}{}}}
\subsection{Method \code{search()}}{
Searches for Person objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$search(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  emailAddresses = NULL,
  firstNames = NULL,
  lastNames = NULL,
  mailingAddresses = NULL,
  middleNames = NULL,
  personDbIds = NULL,
  phoneNumbers = NULL,
  userIDs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{emailAddresses}}{email address for this person}

\item{\code{firstNames}}{Persons first name}

\item{\code{lastNames}}{Persons last name}

\item{\code{mailingAddresses}}{physical address of this person}

\item{\code{middleNames}}{Persons middle name}

\item{\code{personDbIds}}{Unique ID for this person}

\item{\code{phoneNumbers}}{phone number of this person}

\item{\code{userIDs}}{A systems user ID associated with this person.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Person objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-People-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-create"></a>}}
\if{latex}{\out{\hypertarget{method-People-create}{}}}
\subsection{Method \code{create()}}{
Creates a Person object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Person object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-update"></a>}}
\if{latex}{\out{\hypertarget{method-People-update}{}}}
\subsection{Method \code{update()}}{
Creates a Person object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Person object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-People-clone"></a>}}
\if{latex}{\out{\hypertarget{method-People-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{People$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
