% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_PlannedCross.R
\docType{class}
\name{PlannedCrosses}
\alias{PlannedCrosses}
\title{PlannedCrosses Class}
\description{
The PlannedCrosses class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the PlannedCross entity
}
\details{
The PlannedCrosses class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the PlannedCross entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlannedCrosses-new}{\code{PlannedCrosses$new()}}
\item \href{#method-PlannedCrosses-getAll}{\code{PlannedCrosses$getAll()}}
\item \href{#method-PlannedCrosses-create}{\code{PlannedCrosses$create()}}
\item \href{#method-PlannedCrosses-update}{\code{PlannedCrosses$update()}}
\item \href{#method-PlannedCrosses-clone}{\code{PlannedCrosses$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlannedCrosses-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlannedCrosses-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlannedCrosses$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlannedCrosses-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-PlannedCrosses-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list PlannedCross objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlannedCrosses$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  crossingProjectDbIds = NULL,
  crossingProjectNames = NULL,
  plannedCrossDbIds = NULL,
  plannedCrossNames = NULL,
  statuses = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{crossingProjectDbIds}}{Search for Crossing Projects with this unique id}

\item{\code{crossingProjectNames}}{The human readable name for a crossing project}

\item{\code{plannedCrossDbIds}}{Search for Planned Cross with this unique id}

\item{\code{plannedCrossNames}}{Search for Planned Cross with this human readable name}

\item{\code{statuses}}{The status of this planned cross.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of PlannedCross objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlannedCrosses-create"></a>}}
\if{latex}{\out{\hypertarget{method-PlannedCrosses-create}{}}}
\subsection{Method \code{create()}}{
Creates a PlannedCross object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlannedCrosses$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new PlannedCross object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlannedCrosses-update"></a>}}
\if{latex}{\out{\hypertarget{method-PlannedCrosses-update}{}}}
\subsection{Method \code{update()}}{
Creates a PlannedCross object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlannedCrosses$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated PlannedCross object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlannedCrosses-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlannedCrosses-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlannedCrosses$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
