% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Plate.R
\docType{class}
\name{Plates}
\alias{Plates}
\title{Plates Class}
\description{
The Plates class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Plate entity
}
\details{
The Plates class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Plate entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plates-new}{\code{Plates$new()}}
\item \href{#method-Plates-get}{\code{Plates$get()}}
\item \href{#method-Plates-getAll}{\code{Plates$getAll()}}
\item \href{#method-Plates-search}{\code{Plates$search()}}
\item \href{#method-Plates-searchResult}{\code{Plates$searchResult()}}
\item \href{#method-Plates-create}{\code{Plates$create()}}
\item \href{#method-Plates-clone}{\code{Plates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plates-new"></a>}}
\if{latex}{\out{\hypertarget{method-Plates-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plates$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plates-get"></a>}}
\if{latex}{\out{\hypertarget{method-Plates-get}{}}}
\subsection{Method \code{get()}}{
Gets a Plate object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plates$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Plate to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Plate object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plates-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Plates-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Plate objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plates$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationUnitDbIds = NULL,
  plateDbIds = NULL,
  plateNames = NULL,
  plateBarcodes = NULL,
  sampleDbIds = NULL,
  sampleNames = NULL,
  sampleGroupDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationUnitDbIds}}{The ID which uniquely identifies an observation unit}

\item{\code{plateDbIds}}{The ID which uniquely identifies a plate of samples}

\item{\code{plateNames}}{The human readable name of a plate of samples}

\item{\code{plateBarcodes}}{A unique identifier physically attached to the plate}

\item{\code{sampleDbIds}}{The ID which uniquely identifies a sample}

\item{\code{sampleNames}}{The human readable name of the sample}

\item{\code{sampleGroupDbIds}}{The unique identifier for a group of related Samples}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Plate objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plates-search"></a>}}
\if{latex}{\out{\hypertarget{method-Plates-search}{}}}
\subsection{Method \code{search()}}{
Searches for Plate objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plates$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationUnitDbIds = NULL,
  plateDbIds = NULL,
  plateNames = NULL,
  plateBarcodes = NULL,
  sampleDbIds = NULL,
  sampleNames = NULL,
  sampleGroupDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationUnitDbIds}}{The ID which uniquely identifies an observation unit}

\item{\code{plateDbIds}}{The ID which uniquely identifies a plate of samples}

\item{\code{plateNames}}{The human readable name of a plate of samples}

\item{\code{plateBarcodes}}{A unique identifier physically attached to the plate}

\item{\code{sampleDbIds}}{The ID which uniquely identifies a sample}

\item{\code{sampleNames}}{The human readable name of the sample}

\item{\code{sampleGroupDbIds}}{The unique identifier for a group of related Samples
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Plate objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plates-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Plates-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plates$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plates-create"></a>}}
\if{latex}{\out{\hypertarget{method-Plates-create}{}}}
\subsection{Method \code{create()}}{
Creates a Plate object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plates$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Plate object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plates-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plates-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
