% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Program.R
\docType{class}
\name{Programs}
\alias{Programs}
\title{Programs Class}
\description{
The Programs class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Program entity
}
\details{
The Programs class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Program entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Programs-new}{\code{Programs$new()}}
\item \href{#method-Programs-get}{\code{Programs$get()}}
\item \href{#method-Programs-getAll}{\code{Programs$getAll()}}
\item \href{#method-Programs-search}{\code{Programs$search()}}
\item \href{#method-Programs-searchResult}{\code{Programs$searchResult()}}
\item \href{#method-Programs-create}{\code{Programs$create()}}
\item \href{#method-Programs-update}{\code{Programs$update()}}
\item \href{#method-Programs-clone}{\code{Programs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-new"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-get"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-get}{}}}
\subsection{Method \code{get()}}{
Gets a Program object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Program to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Program object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Program objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  abbreviations = NULL,
  leadPersonDbIds = NULL,
  leadPersonNames = NULL,
  objectives = NULL,
  programTypes = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{abbreviations}}{A list of shortened human readable names for a set of Programs}

\item{\code{leadPersonDbIds}}{The person DbIds of the program leader to search for}

\item{\code{leadPersonNames}}{The names of the program leader to search for}

\item{\code{objectives}}{A program objective to search for}

\item{\code{programTypes}}{The type of program entity this object represents <br/> 'STANDARD' represents a standard, permanent breeding program <br/> 'PROJECT' represents a short term project, usually with a set time limit based on funding}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Program objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-search"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-search}{}}}
\subsection{Method \code{search()}}{
Searches for Program objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$search(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  abbreviations = NULL,
  leadPersonDbIds = NULL,
  leadPersonNames = NULL,
  objectives = NULL,
  programTypes = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{abbreviations}}{A list of shortened human readable names for a set of Programs}

\item{\code{leadPersonDbIds}}{The person DbIds of the program leader to search for}

\item{\code{leadPersonNames}}{The names of the program leader to search for}

\item{\code{objectives}}{A program objective to search for}

\item{\code{programTypes}}{The type of program entity this object represents <br/> 'STANDARD' represents a standard, permanent breeding program <br/> 'PROJECT' represents a short term project, usually with a set time limit based on funding
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Program objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-create"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-create}{}}}
\subsection{Method \code{create()}}{
Creates a Program object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Program object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-update"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-update}{}}}
\subsection{Method \code{update()}}{
Creates a Program object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Program object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Programs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Programs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Programs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
