% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Reference.R
\docType{class}
\name{References}
\alias{References}
\title{References Class}
\description{
The References class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Reference entity
}
\details{
The References class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Reference entity
}
\examples{
References$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-References-new}{\code{References$new()}}
\item \href{#method-References-get}{\code{References$get()}}
\item \href{#method-References-getAll}{\code{References$getAll()}}
\item \href{#method-References-search}{\code{References$search()}}
\item \href{#method-References-searchResult}{\code{References$searchResult()}}
\item \href{#method-References-clone}{\code{References$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-References-new"></a>}}
\if{latex}{\out{\hypertarget{method-References-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{References$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-References-get"></a>}}
\if{latex}{\out{\hypertarget{method-References-get}{}}}
\subsection{Method \code{get()}}{
Gets a Reference object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{References$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Reference to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Reference object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-References-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-References-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Reference objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{References$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  accessions = NULL,
  md5checksums = NULL,
  referenceDbIds = NULL,
  referenceSetDbIds = NULL,
  isDerived = NULL,
  minLength = NULL,
  maxLength = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{accessions}}{If specified, return the references for which the \code{accession} matches this string (case-sensitive, exact match).}

\item{\code{md5checksums}}{If specified, return the references for which the \code{md5checksum} matches this string (case-sensitive, exact match).}

\item{\code{referenceDbIds}}{A list of IDs which uniquely identify \code{References} within the given database server}

\item{\code{referenceSetDbIds}}{A list of IDs which uniquely identify \code{ReferenceSets} within the given database server}

\item{\code{isDerived}}{A sequence X is said to be derived from source sequence Y, if X and Y are of the same length and the per-base sequence divergence at A/C/G/T bases is sufficiently small.}

\item{\code{minLength}}{The minimum length of this \code{References} sequence.}

\item{\code{maxLength}}{The minimum length of this \code{References} sequence.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Reference objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-References-search"></a>}}
\if{latex}{\out{\hypertarget{method-References-search}{}}}
\subsection{Method \code{search()}}{
Searches for Reference objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{References$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  accessions = NULL,
  md5checksums = NULL,
  referenceDbIds = NULL,
  referenceSetDbIds = NULL,
  isDerived = NULL,
  minLength = NULL,
  maxLength = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{accessions}}{If specified, return the references for which the \code{accession} matches this string (case-sensitive, exact match).}

\item{\code{md5checksums}}{If specified, return the references for which the \code{md5checksum} matches this string (case-sensitive, exact match).}

\item{\code{referenceDbIds}}{A list of IDs which uniquely identify \code{References} within the given database server}

\item{\code{referenceSetDbIds}}{A list of IDs which uniquely identify \code{ReferenceSets} within the given database server}

\item{\code{isDerived}}{A sequence X is said to be derived from source sequence Y, if X and Y are of the same length and the per-base sequence divergence at A/C/G/T bases is sufficiently small.}

\item{\code{minLength}}{The minimum length of this \code{References} sequence.}

\item{\code{maxLength}}{The minimum length of this \code{References} sequence.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Reference objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-References-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-References-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{References$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-References-clone"></a>}}
\if{latex}{\out{\hypertarget{method-References-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{References$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
