% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_ReferenceSet.R
\docType{class}
\name{ReferenceSets}
\alias{ReferenceSets}
\title{ReferenceSets Class}
\description{
The ReferenceSets class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the ReferenceSet entity
}
\details{
The ReferenceSets class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the ReferenceSet entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReferenceSets-new}{\code{ReferenceSets$new()}}
\item \href{#method-ReferenceSets-get}{\code{ReferenceSets$get()}}
\item \href{#method-ReferenceSets-getAll}{\code{ReferenceSets$getAll()}}
\item \href{#method-ReferenceSets-search}{\code{ReferenceSets$search()}}
\item \href{#method-ReferenceSets-searchResult}{\code{ReferenceSets$searchResult()}}
\item \href{#method-ReferenceSets-clone}{\code{ReferenceSets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceSets-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceSets-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceSets$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceSets-get"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceSets-get}{}}}
\subsection{Method \code{get()}}{
Gets a ReferenceSet object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceSets$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the ReferenceSet to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a ReferenceSet object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceSets-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceSets-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list ReferenceSet objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceSets$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  accessions = NULL,
  assemblyPUIs = NULL,
  md5checksums = NULL,
  referenceSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{accessions}}{If set, return the reference sets for which the \code{accession} matches this string (case-sensitive, exact match).}

\item{\code{assemblyPUIs}}{If set, return the reference sets for which the \code{assemblyId} matches this string (case-sensitive, exact match).}

\item{\code{md5checksums}}{If set, return the reference sets for which the \code{md5checksum} matches this string (case-sensitive, exact match).}

\item{\code{referenceSetDbIds}}{The \code{ReferenceSets} to search.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of ReferenceSet objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceSets-search"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceSets-search}{}}}
\subsection{Method \code{search()}}{
Searches for ReferenceSet objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceSets$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  accessions = NULL,
  assemblyPUIs = NULL,
  md5checksums = NULL,
  referenceSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{accessions}}{If set, return the reference sets for which the \code{accession} matches this string (case-sensitive, exact match).}

\item{\code{assemblyPUIs}}{If set, return the reference sets for which the \code{assemblyId} matches this string (case-sensitive, exact match).}

\item{\code{md5checksums}}{If set, return the reference sets for which the \code{md5checksum} matches this string (case-sensitive, exact match).}

\item{\code{referenceSetDbIds}}{The \code{ReferenceSets} to search.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for ReferenceSet objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceSets-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceSets-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceSets$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceSets-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceSets-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceSets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
