% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Sample.R
\docType{class}
\name{Samples}
\alias{Samples}
\title{Samples Class}
\description{
The Samples class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Sample entity
}
\details{
The Samples class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Sample entity
}
\examples{
Samples$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Samples-new}{\code{Samples$new()}}
\item \href{#method-Samples-get}{\code{Samples$get()}}
\item \href{#method-Samples-getAll}{\code{Samples$getAll()}}
\item \href{#method-Samples-search}{\code{Samples$search()}}
\item \href{#method-Samples-searchResult}{\code{Samples$searchResult()}}
\item \href{#method-Samples-create}{\code{Samples$create()}}
\item \href{#method-Samples-update}{\code{Samples$update()}}
\item \href{#method-Samples-clone}{\code{Samples$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-new"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-get"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-get}{}}}
\subsection{Method \code{get()}}{
Gets a Sample object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Sample to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Sample object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Sample objects from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationUnitDbIds = NULL,
  plateDbIds = NULL,
  plateNames = NULL,
  sampleDbIds = NULL,
  sampleNames = NULL,
  sampleGroupDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationUnitDbIds}}{The ID which uniquely identifies an \code{ObservationUnit}}

\item{\code{plateDbIds}}{The ID which uniquely identifies a \code{Plate} of \code{Samples}}

\item{\code{plateNames}}{The human readable name of a \code{Plate} of \code{Samples}}

\item{\code{sampleDbIds}}{The ID which uniquely identifies a \code{Sample}}

\item{\code{sampleNames}}{The human readable name of the \code{Sample}}

\item{\code{sampleGroupDbIds}}{The unique identifier for a group of related \code{Samples}}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Sample objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-search"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-search}{}}}
\subsection{Method \code{search()}}{
Searches for Sample objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationUnitDbIds = NULL,
  plateDbIds = NULL,
  plateNames = NULL,
  sampleDbIds = NULL,
  sampleNames = NULL,
  sampleGroupDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationUnitDbIds}}{The ID which uniquely identifies an \code{ObservationUnit}}

\item{\code{plateDbIds}}{The ID which uniquely identifies a \code{Plate} of \code{Samples}}

\item{\code{plateNames}}{The human readable name of a \code{Plate} of \code{Samples}}

\item{\code{sampleDbIds}}{The ID which uniquely identifies a \code{Sample}}

\item{\code{sampleNames}}{The human readable name of the \code{Sample}}

\item{\code{sampleGroupDbIds}}{The unique identifier for a group of related \code{Samples}
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Sample objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-create"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-create}{}}}
\subsection{Method \code{create()}}{
Creates a Sample object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Sample object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-update"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-update}{}}}
\subsection{Method \code{update()}}{
Creates a Sample object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Sample object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Samples-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Samples-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Samples$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
