% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Scale.R
\docType{class}
\name{Scales}
\alias{Scales}
\title{Scales Class}
\description{
The Scales class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Scale entity
}
\details{
The Scales class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Scale entity
}
\examples{
Scales$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scales-new}{\code{Scales$new()}}
\item \href{#method-Scales-get}{\code{Scales$get()}}
\item \href{#method-Scales-getAll}{\code{Scales$getAll()}}
\item \href{#method-Scales-create}{\code{Scales$create()}}
\item \href{#method-Scales-update}{\code{Scales$update()}}
\item \href{#method-Scales-clone}{\code{Scales$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scales-new"></a>}}
\if{latex}{\out{\hypertarget{method-Scales-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scales$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scales-get"></a>}}
\if{latex}{\out{\hypertarget{method-Scales-get}{}}}
\subsection{Method \code{get()}}{
Gets a Scale object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scales$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Scale to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Scale object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scales-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Scales-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Scale objects from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scales$getAll(
  commonCropNames = NULL,
  ontologyDbIds = NULL,
  programDbIds = NULL,
  programNames = NULL,
  scaleDbIds = NULL,
  observationVariableDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{ontologyDbIds}}{The unique identifier for an ontology definition.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{scaleDbIds}}{The unique identifier for a scale.}

\item{\code{observationVariableDbIds}}{The unique identifier for an observation variable.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Scale objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scales-create"></a>}}
\if{latex}{\out{\hypertarget{method-Scales-create}{}}}
\subsection{Method \code{create()}}{
Creates a Scale object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scales$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Scale object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scales-update"></a>}}
\if{latex}{\out{\hypertarget{method-Scales-update}{}}}
\subsection{Method \code{update()}}{
Creates a Scale object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scales$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Scale object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scales-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scales-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scales$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
