% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Season.R
\docType{class}
\name{Seasons}
\alias{Seasons}
\title{Seasons Class}
\description{
The Seasons class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Season entity
}
\details{
The Seasons class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Season entity
}
\examples{
Seasons$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Seasons-new}{\code{Seasons$new()}}
\item \href{#method-Seasons-get}{\code{Seasons$get()}}
\item \href{#method-Seasons-getAll}{\code{Seasons$getAll()}}
\item \href{#method-Seasons-create}{\code{Seasons$create()}}
\item \href{#method-Seasons-update}{\code{Seasons$update()}}
\item \href{#method-Seasons-clone}{\code{Seasons$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Seasons-new"></a>}}
\if{latex}{\out{\hypertarget{method-Seasons-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Seasons$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Seasons-get"></a>}}
\if{latex}{\out{\hypertarget{method-Seasons-get}{}}}
\subsection{Method \code{get()}}{
Gets a Season object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Seasons$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Season to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Season object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Seasons-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Seasons-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Season objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Seasons$getAll(
  seasonDbIds = NULL,
  seasons = NULL,
  seasonNames = NULL,
  years = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seasonDbIds}}{The unique identifier for a season.}

\item{\code{seasons}}{The term to describe a given season.}

\item{\code{seasonNames}}{The term to describe a given season.}

\item{\code{years}}{The 4 digit year of a season.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Season objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Seasons-create"></a>}}
\if{latex}{\out{\hypertarget{method-Seasons-create}{}}}
\subsection{Method \code{create()}}{
Creates a Season object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Seasons$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Season object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Seasons-update"></a>}}
\if{latex}{\out{\hypertarget{method-Seasons-update}{}}}
\subsection{Method \code{update()}}{
Creates a Season object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Seasons$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Season object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Seasons-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Seasons-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Seasons$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
