% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_SeedLot.R
\docType{class}
\name{SeedLots}
\alias{SeedLots}
\title{SeedLots Class}
\description{
The SeedLots class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the SeedLot entity
}
\details{
The SeedLots class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the SeedLot entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeedLots-new}{\code{SeedLots$new()}}
\item \href{#method-SeedLots-get}{\code{SeedLots$get()}}
\item \href{#method-SeedLots-getAll}{\code{SeedLots$getAll()}}
\item \href{#method-SeedLots-create}{\code{SeedLots$create()}}
\item \href{#method-SeedLots-update}{\code{SeedLots$update()}}
\item \href{#method-SeedLots-clone}{\code{SeedLots$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedLots-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeedLots-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedLots$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedLots-get"></a>}}
\if{latex}{\out{\hypertarget{method-SeedLots-get}{}}}
\subsection{Method \code{get()}}{
Gets a SeedLot object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedLots$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the SeedLot to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a SeedLot object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedLots-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-SeedLots-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list SeedLot objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedLots$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  seedLotDbIds = NULL,
  crossDbIds = NULL,
  crossNames = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{seedLotDbIds}}{Unique id for a seed lot on this server}

\item{\code{crossDbIds}}{Search for Cross with this unique id}

\item{\code{crossNames}}{Search for Cross with this human readable name}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of SeedLot objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedLots-create"></a>}}
\if{latex}{\out{\hypertarget{method-SeedLots-create}{}}}
\subsection{Method \code{create()}}{
Creates a SeedLot object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedLots$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new SeedLot object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedLots-update"></a>}}
\if{latex}{\out{\hypertarget{method-SeedLots-update}{}}}
\subsection{Method \code{update()}}{
Creates a SeedLot object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedLots$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated SeedLot object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedLots-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeedLots-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedLots$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
