% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Study.R
\docType{class}
\name{Studies}
\alias{Studies}
\title{Studies Class}
\description{
The Studies class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Study entity
}
\details{
The Studies class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Study entity
}
\examples{
Studies$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Studies-new}{\code{Studies$new()}}
\item \href{#method-Studies-get}{\code{Studies$get()}}
\item \href{#method-Studies-getAll}{\code{Studies$getAll()}}
\item \href{#method-Studies-search}{\code{Studies$search()}}
\item \href{#method-Studies-searchResult}{\code{Studies$searchResult()}}
\item \href{#method-Studies-create}{\code{Studies$create()}}
\item \href{#method-Studies-update}{\code{Studies$update()}}
\item \href{#method-Studies-clone}{\code{Studies$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-new"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-get"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-get}{}}}
\subsection{Method \code{get()}}{
Gets a Study object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Study to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Study object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Study objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  active = NULL,
  seasonDbIds = NULL,
  studyTypes = NULL,
  studyCodes = NULL,
  studyPUIs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{active}}{A flag to indicate if a Study is currently active and ongoing}

\item{\code{seasonDbIds}}{The ID which uniquely identifies a season}

\item{\code{studyTypes}}{The type of study being performed.}

\item{\code{studyCodes}}{A short human readable code for a study}

\item{\code{studyPUIs}}{Permanent unique identifier associated with study data.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Study objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-search"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-search}{}}}
\subsection{Method \code{search()}}{
Searches for Study objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  active = NULL,
  seasonDbIds = NULL,
  studyTypes = NULL,
  studyCodes = NULL,
  studyPUIs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{active}}{A flag to indicate if a Study is currently active and ongoing}

\item{\code{seasonDbIds}}{The ID which uniquely identifies a season}

\item{\code{studyTypes}}{The type of study being performed.}

\item{\code{studyCodes}}{A short human readable code for a study}

\item{\code{studyPUIs}}{Permanent unique identifier associated with study data.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Study objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-create"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-create}{}}}
\subsection{Method \code{create()}}{
Creates a Study object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Study object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-update"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-update}{}}}
\subsection{Method \code{update()}}{
Creates a Study object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Study object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Studies-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Studies-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Studies$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
