% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Trait.R
\docType{class}
\name{Traits}
\alias{Traits}
\title{Traits Class}
\description{
The Traits class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Trait entity
}
\details{
The Traits class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Trait entity
}
\examples{
Traits$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Traits-new}{\code{Traits$new()}}
\item \href{#method-Traits-get}{\code{Traits$get()}}
\item \href{#method-Traits-getAll}{\code{Traits$getAll()}}
\item \href{#method-Traits-create}{\code{Traits$create()}}
\item \href{#method-Traits-update}{\code{Traits$update()}}
\item \href{#method-Traits-clone}{\code{Traits$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Traits-new"></a>}}
\if{latex}{\out{\hypertarget{method-Traits-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Traits$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Traits-get"></a>}}
\if{latex}{\out{\hypertarget{method-Traits-get}{}}}
\subsection{Method \code{get()}}{
Gets a Trait object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Traits$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Trait to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Trait object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Traits-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Traits-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Trait objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Traits$getAll(
  commonCropNames = NULL,
  ontologyDbIds = NULL,
  programDbIds = NULL,
  programNames = NULL,
  traitDbIds = NULL,
  observationVariableDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{ontologyDbIds}}{The unique identifier for an ontology definition.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{traitDbIds}}{The unique identifier for a trait.}

\item{\code{observationVariableDbIds}}{The unique identifier for an observation variable.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Trait objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Traits-create"></a>}}
\if{latex}{\out{\hypertarget{method-Traits-create}{}}}
\subsection{Method \code{create()}}{
Creates a Trait object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Traits$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Trait object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Traits-update"></a>}}
\if{latex}{\out{\hypertarget{method-Traits-update}{}}}
\subsection{Method \code{update()}}{
Creates a Trait object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Traits$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Trait object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Traits-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Traits-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Traits$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
