% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Trial.R
\docType{class}
\name{Trials}
\alias{Trials}
\title{Trials Class}
\description{
The Trials class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Trial entity
}
\details{
The Trials class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Trial entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Trials-new}{\code{Trials$new()}}
\item \href{#method-Trials-get}{\code{Trials$get()}}
\item \href{#method-Trials-getAll}{\code{Trials$getAll()}}
\item \href{#method-Trials-search}{\code{Trials$search()}}
\item \href{#method-Trials-searchResult}{\code{Trials$searchResult()}}
\item \href{#method-Trials-create}{\code{Trials$create()}}
\item \href{#method-Trials-update}{\code{Trials$update()}}
\item \href{#method-Trials-clone}{\code{Trials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-new"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-get"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-get}{}}}
\subsection{Method \code{get()}}{
Gets a Trial object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Trial to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Trial object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Trial objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$getAll(
  commonCropNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  active = NULL,
  contactDbIds = NULL,
  searchDateRangeStart = NULL,
  searchDateRangeEnd = NULL,
  trialPUIs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{active}}{A flag to indicate if a Trial is currently active and ongoing}

\item{\code{contactDbIds}}{List of contact entities associated with this trial}

\item{\code{searchDateRangeStart}}{The start of the overlapping search date range.}

\item{\code{searchDateRangeEnd}}{The end of the overlapping search date range.}

\item{\code{trialPUIs}}{A permanent identifier for a trial.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Trial objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-search"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-search}{}}}
\subsection{Method \code{search()}}{
Searches for Trial objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$search(
  commonCropNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  observationVariableDbIds = NULL,
  observationVariableNames = NULL,
  observationVariablePUIs = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  active = NULL,
  contactDbIds = NULL,
  searchDateRangeStart = NULL,
  searchDateRangeEnd = NULL,
  trialPUIs = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{observationVariableDbIds}}{The DbIds of Variables to search for}

\item{\code{observationVariableNames}}{The names of Variables to search for}

\item{\code{observationVariablePUIs}}{The Permanent Unique Identifier of an Observation Variable, usually in the form of a URI}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{active}}{A flag to indicate if a Trial is currently active and ongoing}

\item{\code{contactDbIds}}{List of contact entities associated with this trial}

\item{\code{searchDateRangeStart}}{The start of the overlapping search date range.}

\item{\code{searchDateRangeEnd}}{The end of the overlapping search date range.}

\item{\code{trialPUIs}}{A permanent identifier for a trial.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Trial objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-create"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-create}{}}}
\subsection{Method \code{create()}}{
Creates a Trial object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Trial object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-update"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-update}{}}}
\subsection{Method \code{update()}}{
Creates a Trial object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Trial object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Trials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Trials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Trials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
