% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Variant.R
\docType{class}
\name{Variants}
\alias{Variants}
\title{Variants Class}
\description{
The Variants class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Variant entity
}
\details{
The Variants class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Variant entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Variants-new}{\code{Variants$new()}}
\item \href{#method-Variants-get}{\code{Variants$get()}}
\item \href{#method-Variants-getAll}{\code{Variants$getAll()}}
\item \href{#method-Variants-search}{\code{Variants$search()}}
\item \href{#method-Variants-searchResult}{\code{Variants$searchResult()}}
\item \href{#method-Variants-clone}{\code{Variants$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variants-new"></a>}}
\if{latex}{\out{\hypertarget{method-Variants-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variants$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variants-get"></a>}}
\if{latex}{\out{\hypertarget{method-Variants-get}{}}}
\subsection{Method \code{get()}}{
Gets a Variant object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variants$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Variant to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Variant object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variants-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Variants-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Variant objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variants$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  callSetDbIds = NULL,
  end = NULL,
  referenceDbId = NULL,
  referenceDbIds = NULL,
  referenceSetDbIds = NULL,
  start = NULL,
  variantDbIds = NULL,
  variantSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{callSetDbIds}}{**Deprecated in v2.}

\item{\code{end}}{The end of the window (0-based, exclusive) for which overlapping variants should be returned.}

\item{\code{referenceDbId}}{**Deprecated in v2.}

\item{\code{referenceDbIds}}{The unique identifier representing a genotype \code{Reference}}

\item{\code{referenceSetDbIds}}{The unique identifier representing a genotype \code{ReferenceSet}}

\item{\code{start}}{The beginning of the window (0-based, inclusive) for which overlapping variants should be returned.}

\item{\code{variantDbIds}}{A list of IDs which uniquely identify \code{Variants}}

\item{\code{variantSetDbIds}}{A list of IDs which uniquely identify \code{VariantSets}}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Variant objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variants-search"></a>}}
\if{latex}{\out{\hypertarget{method-Variants-search}{}}}
\subsection{Method \code{search()}}{
Searches for Variant objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variants$search(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  callSetDbIds = NULL,
  end = NULL,
  referenceDbId = NULL,
  referenceDbIds = NULL,
  referenceSetDbIds = NULL,
  start = NULL,
  variantDbIds = NULL,
  variantSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{callSetDbIds}}{**Deprecated in v2.}

\item{\code{end}}{The end of the window (0-based, exclusive) for which overlapping variants should be returned.}

\item{\code{referenceDbId}}{**Deprecated in v2.}

\item{\code{referenceDbIds}}{The unique identifier representing a genotype \code{Reference}}

\item{\code{referenceSetDbIds}}{The unique identifier representing a genotype \code{ReferenceSet}}

\item{\code{start}}{The beginning of the window (0-based, inclusive) for which overlapping variants should be returned.}

\item{\code{variantDbIds}}{A list of IDs which uniquely identify \code{Variants}}

\item{\code{variantSetDbIds}}{A list of IDs which uniquely identify \code{VariantSets}
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Variant objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variants-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Variants-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variants$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Variants-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Variants-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variants$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
