% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_VariantSet.R
\docType{class}
\name{VariantSets}
\alias{VariantSets}
\title{VariantSets Class}
\description{
The VariantSets class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the VariantSet entity
}
\details{
The VariantSets class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the VariantSet entity
}
\examples{
VariantSets$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VariantSets-new}{\code{VariantSets$new()}}
\item \href{#method-VariantSets-get}{\code{VariantSets$get()}}
\item \href{#method-VariantSets-getAll}{\code{VariantSets$getAll()}}
\item \href{#method-VariantSets-search}{\code{VariantSets$search()}}
\item \href{#method-VariantSets-searchResult}{\code{VariantSets$searchResult()}}
\item \href{#method-VariantSets-clone}{\code{VariantSets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantSets-new"></a>}}
\if{latex}{\out{\hypertarget{method-VariantSets-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSets$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantSets-get"></a>}}
\if{latex}{\out{\hypertarget{method-VariantSets-get}{}}}
\subsection{Method \code{get()}}{
Gets a VariantSet object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSets$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the VariantSet to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a VariantSet object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantSets-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-VariantSets-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list VariantSet objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSets$getAll(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  callSetDbIds = NULL,
  variantDbIds = NULL,
  variantSetDbIds = NULL,
  referenceDbIds = NULL,
  referenceSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{callSetDbIds}}{The unique identifier representing a CallSet}

\item{\code{variantDbIds}}{The unique identifier representing a Variant}

\item{\code{variantSetDbIds}}{The unique identifier representing a VariantSet}

\item{\code{referenceDbIds}}{The unique identifier representing a genotype Reference}

\item{\code{referenceSetDbIds}}{The unique identifier representing a genotype ReferenceSet}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of VariantSet objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantSets-search"></a>}}
\if{latex}{\out{\hypertarget{method-VariantSets-search}{}}}
\subsection{Method \code{search()}}{
Searches for VariantSet objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSets$search(
  commonCropNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  callSetDbIds = NULL,
  variantDbIds = NULL,
  variantSetDbIds = NULL,
  referenceDbIds = NULL,
  referenceSetDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{callSetDbIds}}{The unique identifier representing a CallSet}

\item{\code{variantDbIds}}{The unique identifier representing a Variant}

\item{\code{variantSetDbIds}}{The unique identifier representing a VariantSet}

\item{\code{referenceDbIds}}{The unique identifier representing a genotype Reference}

\item{\code{referenceSetDbIds}}{The unique identifier representing a genotype ReferenceSet
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for VariantSet objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantSets-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-VariantSets-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSets$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VariantSets-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VariantSets-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VariantSets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
