% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Event.R
\docType{class}
\name{Events}
\alias{Events}
\title{Events Class}
\description{
The Events class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Event entity
}
\details{
The Events class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Event entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{GenomeMaps}},
\code{\link{Germplasm}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Events-new}{\code{Events$new()}}
\item \href{#method-Events-getAll}{\code{Events$getAll()}}
\item \href{#method-Events-clone}{\code{Events$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Events-new"></a>}}
\if{latex}{\out{\hypertarget{method-Events-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Events$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Events-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Events-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Event objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Events$getAll(
  studyDbIds = NULL,
  studyNames = NULL,
  observationUnitDbIds = NULL,
  eventDbIds = NULL,
  eventTypes = NULL,
  dateRangeStart = NULL,
  dateRangeEnd = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{observationUnitDbIds}}{The ID which uniquely identifies an observation unit.}

\item{\code{eventDbIds}}{Filter based on an Event DbId.}

\item{\code{eventTypes}}{Filter based on an Event Type}

\item{\code{dateRangeStart}}{Filter based on an Event start date.}

\item{\code{dateRangeEnd}}{Filter based on an Event start date.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Event objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Events-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Events-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Events$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
