% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Germplasm.R
\docType{class}
\name{Germplasm}
\alias{Germplasm}
\title{Germplasm Class}
\description{
The Germplasm class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Germplasm entity
}
\details{
The Germplasm class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Germplasm entity
}
\seealso{
Other generated: 
\code{\link{AlleleMatrix}},
\code{\link{BrAPIClient}},
\code{\link{BreedingMethods}},
\code{\link{CallSets}},
\code{\link{Calls}},
\code{\link{Crosses}},
\code{\link{CrossingProjects}},
\code{\link{Events}},
\code{\link{GenomeMaps}},
\code{\link{GermplasmAttributeValues}},
\code{\link{GermplasmAttributes}},
\code{\link{Images}},
\code{\link{Lists}},
\code{\link{Locations}},
\code{\link{MarkerPositions}},
\code{\link{Methods}},
\code{\link{ObservationUnits}},
\code{\link{ObservationVariables}},
\code{\link{Observations}},
\code{\link{Ontologies}},
\code{\link{PedigreeNodes}},
\code{\link{People}},
\code{\link{PlannedCrosses}},
\code{\link{Plates}},
\code{\link{Programs}},
\code{\link{ReferenceSets}},
\code{\link{References}},
\code{\link{Samples}},
\code{\link{Scales}},
\code{\link{Seasons}},
\code{\link{SeedLots}},
\code{\link{Studies}},
\code{\link{Traits}},
\code{\link{Trials}},
\code{\link{VariantSets}},
\code{\link{Variants}}
}
\concept{generated}
\keyword{generated}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Germplasm-new}{\code{Germplasm$new()}}
\item \href{#method-Germplasm-get}{\code{Germplasm$get()}}
\item \href{#method-Germplasm-getAll}{\code{Germplasm$getAll()}}
\item \href{#method-Germplasm-search}{\code{Germplasm$search()}}
\item \href{#method-Germplasm-searchResult}{\code{Germplasm$searchResult()}}
\item \href{#method-Germplasm-create}{\code{Germplasm$create()}}
\item \href{#method-Germplasm-update}{\code{Germplasm$update()}}
\item \href{#method-Germplasm-clone}{\code{Germplasm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-new"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-get"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-get}{}}}
\subsection{Method \code{get()}}{
Gets a Germplasm object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Germplasm to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Germplasm object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Germplasm objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$getAll(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  germplasmPUIs = NULL,
  accessionNumbers = NULL,
  collections = NULL,
  familyCodes = NULL,
  instituteCodes = NULL,
  binomialNames = NULL,
  genus = NULL,
  species = NULL,
  synonyms = NULL,
  parentDbIds = NULL,
  progenyDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{germplasmPUIs}}{List of Permanent Unique Identifiers to identify germplasm}

\item{\code{accessionNumbers}}{A collection of unique identifiers for materials or germplasm within a genebank  MCPD (v2.}

\item{\code{collections}}{A specific panel/collection/population name this germplasm belongs to.}

\item{\code{familyCodes}}{A familyCode representing the family this germplasm belongs to.}

\item{\code{instituteCodes}}{The code for the institute that maintains the material.}

\item{\code{binomialNames}}{List of the full binomial name (scientific name) to identify a germplasm}

\item{\code{genus}}{List of Genus names to identify germplasm}

\item{\code{species}}{List of Species names to identify germplasm}

\item{\code{synonyms}}{List of alternative names or IDs used to reference this germplasm}

\item{\code{parentDbIds}}{Search for Germplasm with these parents}

\item{\code{progenyDbIds}}{Search for Germplasm with these children}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Germplasm objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-search"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-search}{}}}
\subsection{Method \code{search()}}{
Searches for Germplasm objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$search(
  commonCropNames = NULL,
  germplasmDbIds = NULL,
  germplasmNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  studyDbIds = NULL,
  studyNames = NULL,
  trialDbIds = NULL,
  trialNames = NULL,
  germplasmPUIs = NULL,
  accessionNumbers = NULL,
  collections = NULL,
  familyCodes = NULL,
  instituteCodes = NULL,
  binomialNames = NULL,
  genus = NULL,
  species = NULL,
  synonyms = NULL,
  parentDbIds = NULL,
  progenyDbIds = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{germplasmDbIds}}{List of IDs which uniquely identify germplasm to search for}

\item{\code{germplasmNames}}{List of human readable names to identify germplasm to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{studyDbIds}}{List of study identifiers to search for}

\item{\code{studyNames}}{List of study names to filter search results}

\item{\code{trialDbIds}}{The ID which uniquely identifies a trial to search for}

\item{\code{trialNames}}{The human readable name of a trial to search for}

\item{\code{germplasmPUIs}}{List of Permanent Unique Identifiers to identify germplasm}

\item{\code{accessionNumbers}}{A collection of unique identifiers for materials or germplasm within a genebank  MCPD (v2.}

\item{\code{collections}}{A specific panel/collection/population name this germplasm belongs to.}

\item{\code{familyCodes}}{A familyCode representing the family this germplasm belongs to.}

\item{\code{instituteCodes}}{The code for the institute that maintains the material.}

\item{\code{binomialNames}}{List of the full binomial name (scientific name) to identify a germplasm}

\item{\code{genus}}{List of Genus names to identify germplasm}

\item{\code{species}}{List of Species names to identify germplasm}

\item{\code{synonyms}}{List of alternative names or IDs used to reference this germplasm}

\item{\code{parentDbIds}}{Search for Germplasm with these parents}

\item{\code{progenyDbIds}}{Search for Germplasm with these children
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Germplasm objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-create"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-create}{}}}
\subsection{Method \code{create()}}{
Creates a Germplasm object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Germplasm object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-update"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-update}{}}}
\subsection{Method \code{update()}}{
Creates a Germplasm object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Germplasm object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Germplasm-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Germplasm-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Germplasm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
