% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBrAPI.R
\name{getBrAPI}
\alias{getBrAPI}
\title{Creates a BrAPIClient R6 Class object.}
\usage{
getBrAPI(server = NULL, authentication = NULL, verbose = FALSE)
}
\arguments{
\item{server}{The BraPI server URL to be used}

\item{authentication}{The authentication string to be used if the server requires authentication}

\item{verbose}{Set to TRUE to provide more verbose information what is happening}
}
\value{
A configured BrAPIClient R6 Class object.
}
\description{
Creates a BrAPIClient R6 Class object.
}
\examples{
getBrAPI("www.brapiserver.com")
}
