% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generated_Location.R
\docType{class}
\name{Locations}
\alias{Locations}
\title{Locations Class}
\description{
The Locations class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Location entity
}
\details{
The Locations class handles calling the BraAPI server and is a wrapper class around httr2 functionality for
the Location entity
}
\examples{
Locations$new()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Locations-new}{\code{Locations$new()}}
\item \href{#method-Locations-get}{\code{Locations$get()}}
\item \href{#method-Locations-getAll}{\code{Locations$getAll()}}
\item \href{#method-Locations-search}{\code{Locations$search()}}
\item \href{#method-Locations-searchResult}{\code{Locations$searchResult()}}
\item \href{#method-Locations-create}{\code{Locations$create()}}
\item \href{#method-Locations-update}{\code{Locations$update()}}
\item \href{#method-Locations-clone}{\code{Locations$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-new"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
It is not recommended that this object is created separately from the BrAPIClient
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$new(client)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client}}{THe BrAPIClient to be used for server communication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-get"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-get}{}}}
\subsection{Method \code{get()}}{
Gets a Location object by DbId from the BrAPI server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The DbIid of the Location to be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a Location object by ID.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-getAll"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-getAll}{}}}
\subsection{Method \code{getAll()}}{
Gets a list Location objects from the BrAPI server
Note the filtering arguments are all optional and can be combined to filter the results
The value of the filters an be a single value or a vector of values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$getAll(
  commonCropNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  abbreviations = NULL,
  altitudeMin = NULL,
  altitudeMax = NULL,
  countryCodes = NULL,
  countryNames = NULL,
  coordinates = NULL,
  instituteAddresses = NULL,
  instituteNames = NULL,
  locationTypes = NULL,
  parentLocationDbIds = NULL,
  parentLocationNames = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{abbreviations}}{A list of shortened human readable names for a set of Locations}

\item{\code{altitudeMin}}{The minimum altitude to search for}

\item{\code{altitudeMax}}{The maximum altitude to search for}

\item{\code{countryCodes}}{\link{ISO_3166-1_alpha-3}(https://en.}

\item{\code{countryNames}}{The full name of the country to search for}

\item{\code{coordinates}}{}

\item{\code{instituteAddresses}}{The street address of the institute to search for}

\item{\code{instituteNames}}{The name of the institute to search for}

\item{\code{locationTypes}}{The type of location this represents (ex.}

\item{\code{parentLocationDbIds}}{The unique identifier for a Location <br/> The Parent Location defines the encompassing location that this location belongs to.}

\item{\code{parentLocationNames}}{A human readable name for a location <br/> The Parent Location defines the encompassing location that this location belongs to.}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns a paged and filtered list of Location objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-search"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-search}{}}}
\subsection{Method \code{search()}}{
Searches for Location objects on the BrAPI server. The server
may return the paged results or a search result ID for later retrieval
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$search(
  commonCropNames = NULL,
  locationDbIds = NULL,
  locationNames = NULL,
  programDbIds = NULL,
  programNames = NULL,
  abbreviations = NULL,
  altitudeMin = NULL,
  altitudeMax = NULL,
  countryCodes = NULL,
  countryNames = NULL,
  coordinates = NULL,
  instituteAddresses = NULL,
  instituteNames = NULL,
  locationTypes = NULL,
  parentLocationDbIds = NULL,
  parentLocationNames = NULL,
  page = 0,
  pageSize = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commonCropNames}}{The BrAPI Common Crop Name is the simple, generalized, widely accepted name of the organism being researched.}

\item{\code{locationDbIds}}{The location ids to search for}

\item{\code{locationNames}}{A human readable names to search for}

\item{\code{programDbIds}}{A BrAPI Program represents the high level organization or group who is responsible for conducting trials and studies.}

\item{\code{programNames}}{Use this parameter to only return results associated with the given program names.}

\item{\code{abbreviations}}{A list of shortened human readable names for a set of Locations}

\item{\code{altitudeMin}}{The minimum altitude to search for}

\item{\code{altitudeMax}}{The maximum altitude to search for}

\item{\code{countryCodes}}{\link{ISO_3166-1_alpha-3}(https://en.}

\item{\code{countryNames}}{The full name of the country to search for}

\item{\code{coordinates}}{}

\item{\code{instituteAddresses}}{The street address of the institute to search for}

\item{\code{instituteNames}}{The name of the institute to search for}

\item{\code{locationTypes}}{The type of location this represents (ex.}

\item{\code{parentLocationDbIds}}{The unique identifier for a Location <br/> The Parent Location defines the encompassing location that this location belongs to.}

\item{\code{parentLocationNames}}{A human readable name for a location <br/> The Parent Location defines the encompassing location that this location belongs to.
using the searchResult function}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or an ID to the results.

Get the result of a search for Location objects on the BrAPI server. If the server
returned search result ID for later retrieval, use this function to get the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-searchResult"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-searchResult}{}}}
\subsection{Method \code{searchResult()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$searchResult(searchResultId, page = 0, pageSize = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{searchResultId}}{The search result ID returned from a previous search call}

\item{\code{page}}{The page number of results to return, starting from 0}

\item{\code{pageSize}}{The maximum number of results to return per page}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns the search result or by ID to the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-create"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-create}{}}}
\subsection{Method \code{create()}}{
Creates a Location object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$create(newValue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newValue}}{The new Location object to be created}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-update"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-update}{}}}
\subsection{Method \code{update()}}{
Creates a Location object on the BrAPI server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$update(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The updated Location object to be sent to the server}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locations-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Locations-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locations$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
