% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{oauth_flow}
\alias{oauth_flow}
\title{Creates an OAuth provider for BrAPIClient R6 Class object.
'}
\usage{
oauth_flow(
  id,
  secret = NULL,
  token_url,
  name = NULL,
  auth_url = NULL,
  scope = NULL,
  pkce = TRUE,
  auth_params = NULL,
  token_params = NULL,
  redirect_uri = NULL
)
}
\arguments{
\item{id}{The client ID to be used for OAuth Authorization.}

\item{secret}{The client secret to be used for OAuth Authorization.}

\item{token_url}{The token URL to be used for OAuth Authorization.}

\item{name}{The name to be used for the OAuth client.}

\item{auth_url}{The authorization URL to be used for OAuth Authorization.}

\item{scope}{The scope to be used for OAuth Authorization.}

\item{pkce}{Whether to use PKCE (Proof Key for Code Exchange) for OAuth Authorization. Default is TRUE.}

\item{auth_params}{Additional authorization parameters to be used for OAuth Authorization.}

\item{token_params}{Additional token parameters to be used for OAuth Authorization.}

\item{redirect_uri}{The redirect URI to be used for OAuth Authorization.}
}
\value{
A configured OAuth provider R6 Class object.
}
\description{
Creates an OAuth provider for BrAPIClient R6 Class object.
'
}
\examples{
oauth_flow("client_id", "client_secret", "https://example.com/token", "https://example.com/auth")
}
